% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{applyXlsxConditionalFormatByColumn}
\alias{applyXlsxConditionalFormatByColumn}
\title{Apply Xlsx Conditional Formatting by Column}
\usage{
applyXlsxConditionalFormatByColumn(
  file,
  x,
  sheet = 1,
  numColumns = NULL,
  numStyle = c("#F2F2F2", "#6464FC", "#2929D4"),
  mid_rule = c("mid", "mean", "median"),
  dryrun = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Apply Xlsx Conditional Formatting by Column
}
\details{
This function applies conditional formatting to \code{.xlsx}
files, using the numeric range of each column to define
the color rules. It is an extension to
\code{jamba::applyXlsxConditionalFormat()} that differs only
in applying colors to the data range, rather than using
a pre-defined fixed range across all columns.
}
\concept{jam export functions}
