% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ucsctracks.R
\name{assign_track_defaults}
\alias{assign_track_defaults}
\title{Assign default UCSC track values to environment}
\usage{
assign_track_defaults(
  env = new.env(),
  defaults,
  singlet_only = TRUE,
  track_types = c("track", "superTrack", "parent"),
  type_mapping = c(description = c("longLabel")),
  ...
)
}
\arguments{
\item{env}{\code{environment} in which the default values will be
assigned. If the environment already has these values assigned,
those values will be replaced with values in \code{defaults}.}

\item{defaults}{\code{list} or \code{data.frame} of named values,
whose names are used for assignment in the environment \code{env}.}

\item{singlet_only}{logical indicating whether to perform assignment
only when there is one unique value to assign. Useful when passing
\code{defaults} as a \code{data.frame}, when \code{singlet_only=TRUE} it will
only assign columns containing one unique value.}

\item{track_types}{character vector containing regular expression
patterns, used to recognize track name fields. The values for
track name fields are passed to \code{make_ucsc_trackname()} which
replaces invalid characters with underscore.}

\item{...}{additional arguments are passed to \code{base::assign()}.}
}
\value{
`environment`` is returned invisibly.
}
\description{
Assign default UCSC track values to environment
}
\details{
This function assigns values by name, to the specified
environment. The values can be supplied as a list, or
data.frame.

When \code{singlet_only=TRUE}, each named value must have only
one unique value, otherwise it is not assigned to the environment.

The \code{track_types} argument is used to help handle track names
by removing invalid track characters, for example for \code{"track"},
\code{"superTrack"}, and \code{"parent"}, these must be modified consistently
for the track hierarchy to remain valid.
}
\seealso{
Other jam ucsc browser functions:
\code{\link[=get_track_defaults]{get_track_defaults()}},
\code{\link[=make_ucsc_trackname]{make_ucsc_trackname()}},
\code{\link[=parse_ucsc_gokey]{parse_ucsc_gokey()}}
}
\concept{jam ucsc browser functions}
