% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-nmathm-draw.R
\name{draw_nmathm}
\alias{draw_nmathm}
\title{Draw heatmaps using output from nmatlist2heatmaps()}
\usage{
draw_nmathm(
  nmathm,
  hm_nrow = 1,
  do_plot = TRUE,
  title = NULL,
  caption = NULL,
  ...
)
}
\arguments{
\item{nmathm}{\code{list} output from \code{nmatlist2heatmaps()}.}

\item{hm_nrow}{\code{integer} to define the number of rows to display heatmaps,
currently experimental. In general, multi-row heatmap output is not
easy to define as general-use visualization.
Should color keys be shared by all rows?
Should heatmaps break at exactly half the total number of heatmaps?
Should row annotations be displayed on each row of heatmaps? And only
display the color key for the first row of heatmaps?
It is complicated, and not yet well-explored.}

\item{do_plot}{\code{logical} default TRUE, whether to draw the heatmap, or
just prepare the object(s) to be drawn.
Currently experimental, output formats are not stable yet.
This option is mainly a placeholder to be able to run this function
and produce \emph{something} without actually rendering.}

\item{title, caption}{\code{character} with optional overall title and caption.
By default it uses \code{nmathm$metadata$title} and \code{nmathm$metadata$caption}
as already defined, but these options permit custom values.}

\item{...}{additional arguments are passed to internal functions.}
}
\description{
Draw heatmaps using output from nmatlist2heatmaps()
}
