% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import.R
\name{get_numeric_transform}
\alias{get_numeric_transform}
\title{Get appropriate numeric transformation function}
\usage{
get_numeric_transform(transform, ...)
}
\arguments{
\item{transform}{\code{character} string or \code{function}, or \code{list}
that may contain \code{character} string or \code{function}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{function} or \code{NULL} when no matching function is
found, or \code{list} is returned when the input \code{transform}
has multiple values.
}
\description{
Get appropriate numeric transformation function
}
\details{
This function recognizes numeric transformation functions by
name, or searches the attached R package environments for
a matching function by name.

Recognized transform names:
\itemize{
\item \code{"none"} or \code{"linear"} returns the data without change
\item \code{"log2signed"} applies \code{jamba::log2signed()}, defined as
\code{log2(1+x)} transform to the absolute value, then multiplies
by the original \code{sign(x)}
\item \code{"exp2signed"} applies the inverse of \code{"log2signed"}, which
exponentiates numeric values which were previously transformed
with \code{log2(1+x)}. Note that value=0 when exponentiated becomes \code{+1}
\item \code{"sqrt"} applies square root transform, equivalent to \code{sqrt()}
\item \code{"cubert"} applies cube root \code{x^(1/3)}
\item \code{"qrt"} applies fourth root \code{x^(1/4)}
\item \code{"frt"} or \code{"fthrt"} applies fifth root \code{x^(1/5)}
\item \code{"square"} applies \code{x^2} to absolute value, multiplied by the \code{sign(x)}
\item \code{"cube"} applies \code{x^3};
}

Any other character name is used to find a function with the same
name, and if the function is found it is used. The function is
expected to take input \code{x} and return corresponding output with the
same length as the input. A function such as \code{max()} does not fit
this criteria, but a function such as \code{log2()} is acceptable.
}
\examples{
get_numeric_transform("log2signed")

transform_list <- get_numeric_transform(
   c("none",
   "log2signed",
   log2,
   sqrt));
jamba::sdim(transform_list);

x <- jamba::nameVector(0:16);
transform_list[[1]](x)

transform_list[[2]](x)

}
\seealso{
Other jam utility functions:
\code{\link[=cardinality]{cardinality()}},
\code{\link[=color_complement]{color_complement()}},
\code{\link[=convert_PD_df_to_SE]{convert_PD_df_to_SE()}},
\code{\link[=convert_imputed_assays_to_na]{convert_imputed_assays_to_na()}},
\code{\link[=curate_se_colData]{curate_se_colData()}},
\code{\link[=curate_to_df_by_pattern]{curate_to_df_by_pattern()}},
\code{\link[=design2layout]{design2layout()}},
\code{\link[=handle_df_args]{handle_df_args()}},
\code{\link[=merge_proteomics_se]{merge_proteomics_se()}},
\code{\link[=nmat_summary]{nmat_summary()}},
\code{\link[=nmatlist_summary]{nmatlist_summary()}},
\code{\link[=rmd_tab_iterator]{rmd_tab_iterator()}},
\code{\link[=rowNormScale]{rowNormScale()}},
\code{\link[=summit_from_vector]{summit_from_vector()}}
}
\concept{jam utility functions}
