% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{get_salmon_meta}
\alias{get_salmon_meta}
\title{Get Salmon metadata and aux info into a data.frame}
\usage{
get_salmon_meta(metafile, exclude_hashes = TRUE, ...)
}
\arguments{
\item{metafile}{\code{character} vector of one or more files, usually the
full file path to the \code{meta_info.json} file after running Salmon
quant. The path \code{metafile} should be the path to any output file
from one Salmon quant analysis.}

\item{exclude_hashes}{\code{logical} indicating whether to drop columns that
contain file hashes.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} whose number of rows is equal to the number
of unique Salmon root directories in the input \code{metafile}.
For any input \code{metafile} not found, the output is \code{NULL}.
}
\description{
Get Salmon metadata and aux info into a data.frame
}
\details{
This function takes a file path to one or more Salmon
output files, uses that path to locate the full set of
available files, loads data from each of the discovered
files, and returns the results in a \code{data.frame} format.

This function uses \code{rprojroot::find_root()} to find the
root directory, defined as the directory that contains
the file \code{"cmd_info.json"}. The path to \code{"meta_info.json"}
is constructed relative to that location.

Recognized files:
\itemize{
\item \code{meta_info.json} - typically in a subdirectory \code{aux_info/meta_info.json}
\item \code{cmd_info.json} - typically in the same directory as the \code{aux_info}
directory.
}

If a relative path to \code{"cmd_info.json"} cannot be determined, this
function returns \code{NULL}.

When the input \code{metafile} includes multiple files, only
the unique Salmon root directories are returned.

This function uses \code{jsonlite} to read each JSON file, which
is converted to a \code{data.frame}. Any JSON fields that contain
multiple values are comma-delimited using \code{jamba::cPaste()}
in order to fit on one row in the \code{data.frame}.
}
\examples{
cmdinfopath <- system.file("data", "salmonOut", "cmd_info.json", package="platjam");
if (nchar(cmdinfopath) > 0) {
   get_salmon_meta(cmdinfopath);
}

}
\seealso{
Other jam nextgen sequence functions:
\code{\link[=get_salmon_root]{get_salmon_root()}},
\code{\link[=parse_salmon_flenfile]{parse_salmon_flenfile()}},
\code{\link[=save_salmon_qc_xlsx]{save_salmon_qc_xlsx()}}
}
\concept{jam nextgen sequence functions}
