% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-handle-df-args.R
\name{handle_df_args}
\alias{handle_df_args}
\title{Handle function arguments passed in the form of a data.frame}
\usage{
handle_df_args(
  df,
  exclude = NULL,
  parent_fn = NULL,
  fuzzy_match = TRUE,
  fn_env = NULL,
  ...
)
}
\arguments{
\item{df}{\code{data.frame} with \code{colnames()} that should be applied to
function arguments defined by \code{parent_fn}; or a \code{character}
filename that contains delimited data that can be coerced to
a \code{data.frame}, for example a tab-delimited or comma-delimited
text file.}

\item{parent_fn}{\code{function} used to obtain the \code{formals()} to define
function arguments.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{environment} containing variables parsed from \code{df}
}
\description{
Handle function arguments passed in the form of a data.frame
}
\details{
This function is an internal utility function intended to allow
funtion arguments to be passed as a single \code{data.frame}, and
each of the \code{colnames(df)} represent formal function arguments.

One tricky aspect is how to handle arguments that only expect
one value. It could be validated here, or that could be
responsibility of the parent function.
}
\examples{
parent_fn <- nmatlist2heatmaps;

}
\seealso{
Other jam utility functions:
\code{\link[=cardinality]{cardinality()}},
\code{\link[=color_complement]{color_complement()}},
\code{\link[=convert_PD_df_to_SE]{convert_PD_df_to_SE()}},
\code{\link[=convert_imputed_assays_to_na]{convert_imputed_assays_to_na()}},
\code{\link[=curate_se_colData]{curate_se_colData()}},
\code{\link[=curate_to_df_by_pattern]{curate_to_df_by_pattern()}},
\code{\link[=design2layout]{design2layout()}},
\code{\link[=get_numeric_transform]{get_numeric_transform()}},
\code{\link[=merge_proteomics_se]{merge_proteomics_se()}},
\code{\link[=nmat_summary]{nmat_summary()}},
\code{\link[=nmatlist_summary]{nmatlist_summary()}},
\code{\link[=rmd_tab_iterator]{rmd_tab_iterator()}},
\code{\link[=rowNormScale]{rowNormScale()}},
\code{\link[=summit_from_vector]{summit_from_vector()}}
}
\concept{jam utility functions}
