% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-metabolomics-niehs.R
\name{import_metabolomics_niehs}
\alias{import_metabolomics_niehs}
\title{Import metabolomics data from NIEHS file formats}
\usage{
import_metabolomics_niehs(
  data_path,
  shared_samples_only = TRUE,
  filter_sample_type = NULL,
  curation_txt = NULL,
  drop_na_columns = TRUE,
  drop_na_values = c(NA, "NA", "not applicable", "not specified", "none"),
  simplify_singlet_columns = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data_path}{\code{character} path which contains files as named in
Details.}

\item{shared_samples_only}{\code{logical} indicating whether to retain
only those samples which are shared across positive and negative
ionization data files. This step also typically removed extraneous
technical QC samples which may not have been included in both
the positive and negative ionization data.}

\item{filter_sample_type}{\code{character} with optional subset of
sample types to retain. The most useful is
\code{filter_sample_types="sample"} which will retain only biological
samples, and will drop all other samples. The commonly observed
values in the column \code{"CORE_SampleType"}:
\itemize{
\item \code{"sample"}: biological sample
\item \code{"QC_curve"}: calibration curve sample
\item \code{"blank_system"}: negative control blank sample
\item \code{"<NA>"}: empty values where data provided a sample which was
not described in the associated project metadata file. This
typically occurs only for system quality control checks,
such as "SystemSuitability", "wash", "AqX_blank", and "AqX_sample".
}}

\item{curation_txt}{\code{data.frame} passed to \code{curate_se_colData()}
in order to include sample annotations. The default uses
identifiers from \code{colnames(se)} for each \code{SummarizedExperiment}
object.
column headers found in the annotation metadata file.
If \code{curation_txt} is not supplied, then values will be split into
columns by \verb{_} underscore or \code{" "} whitespace characters.}

\item{drop_na_columns}{\code{logivcal} indicating whether to drop columns
in \code{colData} or \code{rowData} when all values are \code{NA}, \code{"NA"},
\code{"not applicable"}, or \code{"none"}, defined in \code{drop_na_values}.

Note that columns with only one non-na value in all fields will
be removed from the \code{colData(se)} and stored as metadata as
a single \code{character} vector accessible via:
\code{se@metadata$colData_values}}

\item{drop_na_values}{\code{character} vector of values considered to
be "na" when \code{drop_na_columns=TRUE}.}

\item{simplify_singlet_columns}{\code{logical} indicating whether to remove
\code{colData(se)} columns with only one value, instead storing the
name and value in metadata as a character vector.
This option is enabled by default, and simplifies the resulting
\code{colData(se)} so that it only includes columns with two or
more unique values.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of \code{SummarizedExperiment} objects, where the \code{list}
is defined by the type of ionization ("df_pos", "df_neg"), and
the type of data ("cleaned") in the data filenames.
Typically the result includes:
\itemize{
\item \code{"df_pos_cleaned"}
\item \code{"df_neg_cleaned"}
}

If only one ionization is provided, only one entry will be returned.

For each \code{SummarizedExperiment} object:
\itemize{
\item \code{rowData} represents metabolite annotations
\item \code{colData} represents sample annotations, optionally including
annotations via a \code{data.frame} supplied as \code{curation_txt}.
\item the slot \code{"metadata"} is a \code{list} with the following:
\itemize{
\item \code{isample_use}: the subset of \code{colnames(se)} for which there
was sample metadata found in the metadata file. Some control
samples may not match the full metadata, and will be ignored
when using \code{isamples_Use}.
\item \code{irows_use}: all \code{rownames(se)} for all measured metabolites.
\item \code{irows_clean}: the \code{rownames(se)} for measurement with no
annotation in the column \code{"flag_guidance"}.
\item \code{irows_flagged}: the \code{rownames(se)} for measurements with
some non-empty annotation in the column \code{"flag_guidance"}.
}
}
}
\description{
Import metabolomics data from NIEHS file formats
}
\details{
This import function is specific to NIEHS file formats produced
from their defined analysis workflow. The files typically include
\code{"df_pos"} for positive ionization, and \code{"df_neg"} for negative
ionization.

Optionally, when the full data processing file is present,
it will be imported alongside the cleaned data described above.
The full data processing imports detailed compound measurement
data, and is expected in one of two formats in the \code{data_path} folder:
\enumerate{
\item Files \code{"compounds_pos.txt"} and/or \code{"compounds_neg.txt"}, or
\item \code{"1_DataProcessed.zip"} which is expected to contain
files \code{"compounds_pos.txt"} and/or \code{"compounds_neg.txt"} in the archive.
}

The "compounds" data includes important annotations for each measurement,
specifically the type of numeric measurement that is supplied
by the upstream software. These annotations include whether numeric
values were imputed, or measured directly in each sample.
\subsection{Sample Metadata}{
\itemize{
\item \code{"[project_code]_NIEHS_MCF_metadata.txt"}: tab-delimited text file
which contains sample annotations.
}
}

\subsection{Positive Ionization Files}{
\itemize{
\item \code{"df_pos.datamatrix.cleaned.txt"}: Tab-delimited text file
containing peak areas.
Features are processed and cleaned by MCF for quality.
\item \code{"df_pos.datamatrix.cleaned.log10.txt"}: As above, but log10 transformed.
\item \code{"df_pos.datamatrix.cleaned.rowsum.txt"}: As above but using
row sum peak areas.
\item \code{"df_pos.annotation.cleaned.txt"}: annotation of each measured metabolite.
}
}

\subsection{Negative Ionization Files}{
\itemize{
\item \code{"df_neg.datamatrix.cleaned.txt"}: Tab-delimited text file
containing peak areas.
Features are processed and cleaned by MCF for quality.
\item \code{"df_neg.datamatrix.cleaned.log10.txt"}: As above, but log10 transformed.
\item \code{"df_neg.datamatrix.cleaned.rowsum.txt"}: As above but using
row sum peak areas.
\item \code{"df_neg.annotation.cleaned.txt"}: annotation of each measured metabolite.
}
}
}
\examples{
data_path <- path.expand("~/Projects/Rider/metabolomics_jul2023/data");
se_list <- import_metabolomics_niehs(data_path);

}
\seealso{
Other jam import functions:
\code{\link[=coverage_matrix2nmat]{coverage_matrix2nmat()}},
\code{\link[=deepTools_matrix2nmat]{deepTools_matrix2nmat()}},
\code{\link[=frequency_matrix2nmat]{frequency_matrix2nmat()}},
\code{\link[=import_lipotype_csv]{import_lipotype_csv()}},
\code{\link[=import_nanostring_csv]{import_nanostring_csv()}},
\code{\link[=import_nanostring_rcc]{import_nanostring_rcc()}},
\code{\link[=import_nanostring_rlf]{import_nanostring_rlf()}},
\code{\link[=import_omics_data]{import_omics_data()}},
\code{\link[=import_proteomics_PD]{import_proteomics_PD()}},
\code{\link[=import_proteomics_mascot]{import_proteomics_mascot()}},
\code{\link[=import_salmon_quant]{import_salmon_quant()}},
\code{\link[=process_metab_compounds_file]{process_metab_compounds_file()}}
}
\concept{jam import functions}
