% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-mascot.R
\name{import_proteomics_mascot}
\alias{import_proteomics_mascot}
\title{Import proteomics data from Mascot}
\usage{
import_proteomics_mascot(
  file,
  sheet = 1,
  ann_lib = c("org.Hs.eg.db", "org.Mm.eg.db", "org.Rn.eg.db"),
  curation_txt = NULL,
  accession_from = NULL,
  accession_to = NULL,
  xref_df = NULL,
  measurements = c("totalIntensity", "numSpectra"),
  accession_colname = "accession",
  delim = "[/]",
  try_list = c("SYMBOL2EG", "ACCNUM2EG", "UNIPROT2EG", "ENSEMBLPROT2EG", "ALIAS2EG"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{file}{\code{character} path to a file containing proteomics data}

\item{sheet}{\code{integer} or \code{character} name of worksheet when \code{file}
is an Excel \code{xlsx} formatted file.}

\item{ann_lib}{\code{character} passed to \code{genejam::freshenGenes3()}, see
documentation for alternate methods of passing one or more annotation
libraries.}

\item{curation_txt}{\code{data.frame} whose first column should match the
sample column headers found in the PD abundance columns, and
subsequent columns contain associated sample annotations.
If \code{curation_txt} is not supplied, then values will be split into
columns by \verb{_} underscore or \code{" "} whitespace characters.}

\item{accession_from, accession_to}{\code{character} vectors, that help manual
curation from one accession number to another, intended when an
accession number is not recognized by the Bioconductor annotation
library, and a newer accession would be recognized. No gene left
behind.}

\item{xref_df}{\code{data.frame} that contains accession numbers in the
first column, and annotation columns in additional columns, specifically
using \verb{"SYMBOL", "ENTREZID", "GENENAME"} as replacements for
output from \code{genejam::freshenGenes3()}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{jamba::readOpenxlsx()}.}
}
\value{
\code{SummarizedExperiment} object
}
\description{
Import proteomics data from Mascot
}
\examples{
# TODO: replace with smaller test data in the appropriate format
mascot_file <- file.path(path.expand("~/Projects/Hu/hu_msprot_turboid"),
   "Lackford_all_090822.xlsx")
protein_df <- jamba::readOpenxlsx(mascot_file, sheet=1)[[1]];
se <- import_proteomics_mascot;

}
\seealso{
Other jam import functions:
\code{\link[=coverage_matrix2nmat]{coverage_matrix2nmat()}},
\code{\link[=deepTools_matrix2nmat]{deepTools_matrix2nmat()}},
\code{\link[=frequency_matrix2nmat]{frequency_matrix2nmat()}},
\code{\link[=import_lipotype_csv]{import_lipotype_csv()}},
\code{\link[=import_metabolomics_niehs]{import_metabolomics_niehs()}},
\code{\link[=import_nanostring_csv]{import_nanostring_csv()}},
\code{\link[=import_nanostring_rcc]{import_nanostring_rcc()}},
\code{\link[=import_nanostring_rlf]{import_nanostring_rlf()}},
\code{\link[=import_omics_data]{import_omics_data()}},
\code{\link[=import_proteomics_PD]{import_proteomics_PD()}},
\code{\link[=import_salmon_quant]{import_salmon_quant()}},
\code{\link[=process_metab_compounds_file]{process_metab_compounds_file()}}
}
\concept{jam import functions}
