% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-salmonquant.R
\name{import_salmon_quant}
\alias{import_salmon_quant}
\title{Import Salmon quant.sf files to SummarizedExperiment}
\usage{
import_salmon_quant(
  salmonOut_paths,
  import_types = c("tx", "gene", "gene_body", "gene_tx"),
  gtf = NULL,
  tx2gene = NULL,
  curation_txt = NULL,
  tx_colname = "transcript_id",
  gene_colname = "gene_name",
  gene_body_colname = "transcript_type",
  geneFeatureType = "exon",
  txFeatureType = "exon",
  countsFromAbundance = "lengthScaledTPM",
  gene_body_ids = NULL,
  trim_tx_from = NULL,
  trim_tx_to = NULL,
  debug = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{salmonOut_paths}{\code{character} vectors to each individual folder
that contains the \code{"quant.sf"} output file for Salmon.}

\item{import_types}{\code{character} indicating which type or types of
data to return. Note that the distinction between \code{gene} and
\code{gene_body} is only relevant when there are transcript entries
defined with \code{transcript_type="gene_body"}. These entries specifically
represent unspliced transcribed regions for a gene locus, and
only for multi-exon genes.
\itemize{
\item \code{tx}: transcript quantitation, direct import of \code{quant.sf} files.
\item \code{gene}: gene quantitation after calling \code{tximport::summarizeToGene()},
excluding \code{transcript_type="gene_body"}.
\item \code{gene_body}: gene quantitation after calling \code{tximport::summarizeToGene()},
including \code{transcript_type="gene_body"}.
}}

\item{gtf}{\code{character} path to a GTF file, used only when \code{tx2gene}
is not supplied. When used, \code{splicejam::makeTx2geneFromGtf()} is
called to create a \code{data.frame} object \code{tx2gene}.}

\item{tx2gene}{\code{character} path to file, or \code{data.frame} with at
least two columns matching \code{tx_colname} and \code{gene_colname} below.
When supplied, the \code{gtf} argument is ignored, unless the file
path is not accessible, or the data is not \code{data.frame}.}

\item{curation_txt}{\code{data.frame} whose first column should match the
sample column headers found in the PD abundance columns, and
subsequent columns contain associated sample annotations.
If \code{curation_txt} is not supplied, then values will be split into
columns by \verb{_} underscore or \code{" "} whitespace characters.}

\item{tx_colname, gene_colname}{\code{character} strings indicating colnames
in \code{tx2gene} that should be used.
\itemize{
\item \code{tx_colname} represents unique identifier for each transcript,
usually \code{"transcript_id"}.
\item \code{gene_colname} represents a gene label associated with gene
summarized expression values, typically \code{"gene_name"}.
}}

\item{geneFeatureType, txFeatureType}{\code{character} arguments passed to
\code{splicejam::makeTx2geneFromGtf()} only when supplying argument
\code{gtf} with a path to a GTF file.}

\item{countsFromAbundance}{\code{character} string passed to
\code{tximport::summarizeToGene()} to define the method for calculating
abundance.}

\item{gene_body_ids}{\code{character} optional vector with specific row
identifiers that should be considered \code{transcript_type="gene_body"}
entries, relevant to argument \code{import_types} above. When \code{gene_body_ids}
is defined, these entries are used directly without using \code{tx2gene}.
When \code{gene_body_ids} is not defined, \code{tx2gene$transcript_type} is used
if present. If that column is not present, or does not contain any
entries with \code{"gene_body"}, then all transcripts are used for
\code{import_types="gene"}, and \code{import_types="gene_body"} is not valid
and therefore is not returned.}

\item{trim_tx_from, trim_tx_to}{\code{character} vector with one or more
regular expression patterns used to curate the values in \code{tx_colname}
prior to assigning them as \code{rownames()}, and back to \code{tx_colname}.
These values are joined to \code{tx2gene[[tx_colname]]} to assign
additional gene annotations.
\itemize{
\item The default as of version 0.0.79.900 is to leave strand information
\code{"(-)"} and \code{"(+)"} without removing it, since in some rare cases a
gene's unspliced transcripts can be present on two strands.
\item The previous default (version <= 0.0.78.900) was to remove
\code{"(-)"} and \code{"(+)"} from the transcript_id \code{tx_colname} column.
}}

\item{debug}{\code{logical} default FALSE, will return the raw 'txiTx' object.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to supporting functions.}
}
\value{
\code{list} with \code{SummarizedExperiment} objects, each of which
contain assay names \verb{c("counts", "abundance", "length)}, where
\code{c("counts", "abundance")} are transformed with \code{log2(1 + x)}.
The transform can be reversed with \code{10^x - 1}.
The \code{SummarizedExperiment} objects by name:
\itemize{
\item \code{"TxSE"}: transcript-level values imported from \code{quant.sf}.
\item \code{"GeneSE"}: gene-level summary values, excluding
\code{"gene_body"} entries.
\item \code{"GeneBodySE"}: gene-level summary values, summarizing unspliced
and spliced \code{"gene_body"} transcripts together for each gene.
\item \code{"GeneTxSE"}: gene-level summary values, where transcripts are
combined to gene level, and \code{"gene_body"} entries are represented
separately, with suffix \code{"_gene_body"} added to the gene name.
}
}
\description{
Import Salmon quant.sf files to SummarizedExperiment
}
\details{
This function is intended to automate the process of importing
a series of \code{quant.sf} files, then generating \code{SummarizedExperiment}
objects at the transcript and gene level. It optionally includes
sample annotation provided as a \code{data.frame} in argument \code{curation_txt}.
It also includes transcript and gene annotations through either
\code{data.frame} from argument \code{tx2gene}, or it derives \code{tx2gene}
from a GTF file from argument \code{gtf}. The GTF file option then calls
\code{splicejam::makeTx2geneFromGtf()}.

This function can optionally process data that includes full length
gene body regions, annotated with \code{"gene_body"}. This option is specific
for Salmon quantitation where the transcripts include full length
gene body for multi-exon genes, for example to measure unspliced
transcript abundance.
\itemize{
\item \code{import_types="gene"} summarizes only the proper transcripts,
excluding \code{"gene_body"} entries.
\item \code{import_types="gene_body"} summarizes all transcript
and full gene entries into one summary transcript abundance.
\item \code{import_types="gene_tx"} summarizes proper transcript to gene level,
and separately represents \code{"gene_body"} entries for comparison.
}

The current recommendation is to use default values for \code{import_types}
which imports all the following types of data:
\enumerate{
\item \code{TxSE} as transcripts per row
\item \code{GeneSE} as genes per row, excluding unspliced \code{gene_body} transcripts.
\item \code{GeneBodySE} as genes per row, summarizing unspliced
and spliced \code{"gene_body"} transcripts together for each gene.
\item \code{GeneTxSE} as spliced and unspliced genes per row, so that
unspliced, spliced, or both can be analyzed together.
}

We typically use \code{GeneBodySE} and define a subset of \code{rownames(GeneBodySE)}
to use only spliced transcripts during analysis. Optionally we may
run \code{limma::diffSplice()} to compare the unspliced:spliced ratio
across experiment groups.
}
\seealso{
Other jam import functions:
\code{\link[=coverage_matrix2nmat]{coverage_matrix2nmat()}},
\code{\link[=deepTools_matrix2nmat]{deepTools_matrix2nmat()}},
\code{\link[=frequency_matrix2nmat]{frequency_matrix2nmat()}},
\code{\link[=import_lipotype_csv]{import_lipotype_csv()}},
\code{\link[=import_metabolomics_niehs]{import_metabolomics_niehs()}},
\code{\link[=import_nanostring_csv]{import_nanostring_csv()}},
\code{\link[=import_nanostring_rcc]{import_nanostring_rcc()}},
\code{\link[=import_nanostring_rlf]{import_nanostring_rlf()}},
\code{\link[=import_omics_data]{import_omics_data()}},
\code{\link[=import_proteomics_PD]{import_proteomics_PD()}},
\code{\link[=import_proteomics_mascot]{import_proteomics_mascot()}},
\code{\link[=process_metab_compounds_file]{process_metab_compounds_file()}}
}
\concept{jam import functions}
