% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-colors.R
\name{make_jam_divergent}
\alias{make_jam_divergent}
\title{Make divergent color gradient}
\usage{
make_jam_divergent(linear1, linear2, lite = TRUE, n = 21, ...)
}
\description{
Make divergent color gradient using jam_linear and jam_divergent.
}
\examples{
jamba::showColors(jam_linear)

jg1 <- make_jam_divergent("royalblue", "orangered")
jamba::showColors(jg1)
showDichromat(jg1)

jg2 <- make_jam_divergent("slateblue", "firebrick", n=21)
jamba::showColors(jg2)
showDichromat(jg2)

jg3 <- make_jam_divergent("slateblue", "firebrick", lite=FALSE, n=21)
jamba::showColors(jg3)
showDichromat(jg3)

jg4 <- make_jam_divergent("Blues", "Reds", lite=TRUE, n=21)
jamba::showColors(c(jg4,
   list(BuRd=jamba::getColorRamp("RdBu_r", n=21))))

jg5 <- make_jam_divergent("inferno", "dodgerblue1", lite=FALSE, n=21, gradientWtFactor=1)
jamba::showColors(jg5)

xseq <- seq(from=-1, to=1, by=0.1);
mseq <- matrix(xseq, ncol=1);
m <- mseq \%*\% t(mseq);
rownames(m) <- seq_len(nrow(m));
colnames(m) <- seq_len(ncol(m));
hm1 <- ComplexHeatmap::Heatmap(m[,1:10],
   cluster_columns=FALSE,
   cluster_rows=FALSE,
   row_names_side="left",
   border=TRUE,
   heatmap_legend_param=list(
      border="grey10",
      at=seq(from=-1, to=1, by=0.25),
      color_bar="discrete"),
   col=jg3[[1]])

hm2 <- ComplexHeatmap::Heatmap(m[21:1,12:21],
   cluster_columns=FALSE,
   cluster_rows=FALSE,
   border=TRUE,
   heatmap_legend_param=list(
      border=TRUE,
      at=seq(from=-1, to=1, by=0.25),
      color_bar="discrete"),
   col=jg2[[1]])
hm1 + hm2

# same as above but showing where to use lens
hm3 <- ComplexHeatmap::Heatmap(m[,1:10],
   cluster_columns=FALSE,
   cluster_rows=FALSE,
   row_names_side="left",
   border=TRUE,
   heatmap_legend_param=list(
      border=TRUE,
      at=seq(from=-1, to=1, by=0.25),
      color_bar="discrete"),
   col=jamba::getColorRamp(jg3[[1]], divergent=TRUE, lens=2))

hm4 <- ComplexHeatmap::Heatmap(m[21:1,12:21],
   cluster_columns=FALSE,
   cluster_rows=FALSE,
   border=TRUE,
   heatmap_legend_param=list(
      border=TRUE,
      at=seq(from=-1, to=1, by=0.25),
      color_bar="discrete"),
   col=jamba::getColorRamp(jg2[[1]], divergent=TRUE, lens=2))
hm3 + hm4

}
