% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-numeric-to-colorfn.R
\name{mean_hue}
\alias{mean_hue}
\title{Mean hue for a vector}
\usage{
mean_hue(x, seed = 1, ...)
}
\arguments{
\item{x}{\code{numeric} angles in degrees representing HCL color hue.}

\item{seed}{\code{numeric} used with \code{set.seed(seed)} for reproducibility.}

\item{...}{additional arguments are ignored.}
}
\description{
Mean hue for a vector
}
\seealso{
Other jam color functions:
\code{\link[=assign_numeric_colors]{assign_numeric_colors()}},
\code{\link[=design2colors]{design2colors()}},
\code{\link[=df_to_numcolors]{df_to_numcolors()}},
\code{\link[=print_color_list]{print_color_list()}},
\code{\link[=quick_complement_color]{quick_complement_color()}}
}
\concept{jam color functions}
