% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-metabolomics-niehs.R
\name{process_metab_compounds_file}
\alias{process_metab_compounds_file}
\title{Process NIEHS compounds output files (internal)}
\usage{
process_metab_compounds_file(
  x,
  nsmall_mz = 4,
  nsmall_rt = 2,
  assay_names_extra = c("Area", "Gap Fill Status", "Peak Rating"),
  mz_colname = "m/z",
  rt_colname = "RT [min]",
  type = c("pos", "neg"),
  se_list = NULL,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} after reading \code{compound_pos.txt} or \code{compound_neg.txt}}

\item{nsmall_mz}{\code{numeric} argument passed to \code{format()} to define
the number of decimal positions to retain in the output label
from column \code{"m/z"}.}

\item{nsmall_rt}{\code{numeric} argument passed to \code{format()} to define
the number of decimal positions to retain in the output label
from column \code{"RT [min]"}.}

\item{assay_names_extra}{\code{character} vector with prefix values expected
in column names, for example where \code{"Area"} is expected to form a prefix
to column names: \verb{"Area: sample_a", "Area: sample_b"};
and \code{"Gap Fill Status"} would form column names
\verb{"Gap Fill Status: sample_a", "Gap Fill Status: sample_b"}.}

\item{mz_colname, rt_colname}{\code{character} string to indicate the appropriate
column name containing \code{"m/z"} (mass/charge) and \code{"RT [min]"}
(retention time in minutes) as used to format a unique identifier
for each row.}

\item{type}{\code{character} string indicating whether data represents
positive \code{"pos"} or negative \code{"neg"} ionization. Mainly used
when \code{se_list} is also provided.}

\item{se_list}{optional \code{list} of \code{SummarizedExperiment} objects,
named with \code{"pos"} or \code{"neg"} in the name. When provided, the
\code{se} output of this function is merged into the \code{se_list} with
matching \code{type}, adding \code{assay_names} into the corresponding
\code{se_list} object.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{SummarizedExperiment} object with \code{assays} defined by matching
column names in \code{x}. When no column names meet the \code{assay_names_extra}
criteria, this function returns \code{NULL}.
}
\description{
Process NIEHS compounds output files (internal)
}
\details{
This function produces \code{SummarizedExperiment} object with \code{assays}
defined by the \code{assay_names_extra} patterns below. Each assay name
is the lowercase value, with spaces replaced with underscore \code{"_"}.
For example \code{"Gap Fill Status"} will become assay name \code{"gap_fill_status"}.
}
\seealso{
Other jam import functions:
\code{\link[=coverage_matrix2nmat]{coverage_matrix2nmat()}},
\code{\link[=deepTools_matrix2nmat]{deepTools_matrix2nmat()}},
\code{\link[=frequency_matrix2nmat]{frequency_matrix2nmat()}},
\code{\link[=import_lipotype_csv]{import_lipotype_csv()}},
\code{\link[=import_metabolomics_niehs]{import_metabolomics_niehs()}},
\code{\link[=import_nanostring_csv]{import_nanostring_csv()}},
\code{\link[=import_nanostring_rcc]{import_nanostring_rcc()}},
\code{\link[=import_nanostring_rlf]{import_nanostring_rlf()}},
\code{\link[=import_omics_data]{import_omics_data()}},
\code{\link[=import_proteomics_PD]{import_proteomics_PD()}},
\code{\link[=import_proteomics_mascot]{import_proteomics_mascot()}},
\code{\link[=import_salmon_quant]{import_salmon_quant()}}
}
\concept{jam import functions}
