% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-geomx.R
\name{revcomp}
\alias{revcomp}
\title{Simple DNA reverse-complement function}
\usage{
revcomp(x, molecule = c("DNA", "RNA"), ...)
}
\arguments{
\item{x}{\code{character} string with DNA or RNA sequence, expected to
contain only A,T,G,C,N, and possibly U.}

\item{molecule}{\code{character} string with \code{"DNA"} (default) or \code{"RNA"}.
Input recognizes either T or U, but will only produce T for DNA,
and U for RNA.}
}
\value{
\code{character} string with reverse-complemented sequences.
}
\description{
Simple DNA reverse-complement function
}
\examples{
dna_in <- "GTTACTCT";
revcomp(dna_in)

}
\seealso{
Other jam GeoMx functions:
\code{\link[=import_geomx_samplesheet]{import_geomx_samplesheet()}}
}
\concept{jam GeoMx functions}
