% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-neuro.R
\name{rowNormScale}
\alias{rowNormScale}
\title{Normalize and scale data per row}
\usage{
rowNormScale(
  x,
  from = 0,
  to = 1,
  naValue = NA,
  low = NULL,
  high = NULL,
  singletMethod = "min",
  col_range = NULL,
  ...
)
}
\arguments{
\item{x, from, to, naValue, singletMethod}{arguments passed to
\code{jamba::normScale()}. Note that the default \verb{low,high} values
use the column range defined by \code{col_range}.}

\item{low}{numeric value or \code{NULL}, passed to \code{jamba::normScale()}
for each row in \code{x}. When \code{low} is \code{NULL}, it uses the
minimum value in \code{col_range} for each row.}

\item{high}{numeric value or \code{NULL}, passed to \code{jamba::normScale()}
for each row in \code{x}. When \code{high} is \code{NULL}, it uses the
maximum value in \code{col_range} for each row.}

\item{col_range}{integer vector referring to column numbers in
the input \code{x} matrix to use in \code{jamba::normScale()}. When
\code{col_range} is \code{NULL}, it uses all columns in \code{x}.}

\item{...}{additional arguments are passed to \code{jamba::normScale()}.}
}
\description{
Normalize and scale data per row
}
\details{
This function essentially calls \code{jamba::normScale()}
on each row of a numeric matrix. By default, it scales
values to a fixed numeric range from 0 to 1, where
the minimum value is set to 0 and the maximum value is
set to 1. It is much more configurable, see \code{jamba::normScale()}
help docs.
}
\examples{
m <- matrix(1:9, ncol=3);
m;

rowNormScale(m);
rowNormScale(m, from=0, to=10);

}
\seealso{
Other jam utility functions:
\code{\link[=cardinality]{cardinality()}},
\code{\link[=color_complement]{color_complement()}},
\code{\link[=convert_PD_df_to_SE]{convert_PD_df_to_SE()}},
\code{\link[=convert_imputed_assays_to_na]{convert_imputed_assays_to_na()}},
\code{\link[=curate_se_colData]{curate_se_colData()}},
\code{\link[=curate_to_df_by_pattern]{curate_to_df_by_pattern()}},
\code{\link[=design2layout]{design2layout()}},
\code{\link[=get_numeric_transform]{get_numeric_transform()}},
\code{\link[=handle_df_args]{handle_df_args()}},
\code{\link[=merge_proteomics_se]{merge_proteomics_se()}},
\code{\link[=nmat_summary]{nmat_summary()}},
\code{\link[=nmatlist_summary]{nmatlist_summary()}},
\code{\link[=rmd_tab_iterator]{rmd_tab_iterator()}},
\code{\link[=summit_from_vector]{summit_from_vector()}}
}
\concept{jam utility functions}
