% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-contrasts.R
\name{run_contrasts_se}
\alias{run_contrasts_se}
\title{Run contrasts on SummarizedExperiment data}
\usage{
run_contrasts_se(
  se,
  signalList = head(names(assays(se)), 1),
  cutoff_p = NULL,
  cutoff_adjp = 0.05,
  cutoff_fold = 1.5,
  cutoff_mgm = 0,
  cutoff_p_int = cutoff_p,
  cutoff_adjp_int = cutoff_adjp,
  cutoff_fold_int = cutoff_fold,
  confint = FALSE,
  probe_colname = NULL,
  gene_colname = "gene_name",
  idesign = NULL,
  icontrasts = NULL,
  enforce_design_subset = TRUE,
  igenes = NULL,
  isamples = NULL,
  use_voom = FALSE,
  weights = NULL,
  robust = FALSE,
  apply_probe_replicates = FALSE,
  floor_minimum = 0,
  floor_value = 0,
  fill_na = FALSE,
  na_weight = 0,
  return_lmfit = FALSE,
  return_df = TRUE,
  return_dfs = TRUE,
  return_hit_array = TRUE,
  collapse_by_gene = FALSE,
  stat_method = c("limma", "NanoStringDiff"),
  debug = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object with genes/proteins/probes
in each row, and biological samples in each column.}

\item{cutoff_p, cutoff_adjp, cutoff_fold, cutoff_mgm}{numeric vectors or
\code{NULL} to define the thresholds to apply when defining statistical
hits. When multiple values are supplied, each vector is recycled
to the max length of these vectors. The hit criteria are applied
for the first element in each vector, then the second element in
each vector, and so on.}

\item{cutoff_p_int, cutoff_adjp_int, cutoff_fold_int}{numeric vectors
applied only to two-way or higher contrasts. For example, one
may want to apply \code{cutoff_adjp=0.01} for high stringency for
simple pairwise contrasts, then \code{cutoff_adjp_int=0.05} for
less stringent criteria for two-way contrasts, due to the decrease
in statistical power when testing twice as many groups.}

\item{confint}{\code{logical} passed to \code{limma::topTable()} to indicate
whether it returns upper and lower confidence intervals.}

\item{probe_colname, gene_colname}{\code{character} values indicating
the colname in \code{rowData(se)} that contains probe or gene information.
Probe values are expected to represent unique probe assays,
or unique transcript identified, or some equivalent definition.
Gene values represent the gene or protein to which each individual
entry is assigned, and is optional except when \code{collapse_by_gene=TRUE}.}

\item{idesign, icontrasts}{\verb{numeric matrix} to define the design and
statistical contrasts. It is recommended that \code{idesign} has
rownames that match \code{colnames(se)}; and \code{colnames(idesign)} should
match \code{rownames(contrasts)}. The \code{colnames(idesign)} should not
contain characters that interfere with contrast definition, for
example no \code{"-"}, \code{"("}, \code{")"} characters. The \code{colnames(icontrasts)}
can actually contain these characters, as this function makes
effort to maintain the actual \code{colnames(icontrasts)} in the
output for data integrity.}

\item{enforce_design_subset}{\code{logical} indicating whether to use only
the subset of samples represented in the full set of contrasts
provided by \code{icontrasts}. For example, if testing only one two-group
contrast from data that contains four groups,
\code{enforce_design_subset=TRUE} will first reduce the input data to
those two groups, then will run the limma workflow;
\code{enforce_design_subset=FALSE} will keep all four groups, but only
apply the contrast to the two groups. The latter process will use
all four groups in the estimate of variance, which is applied to
the two groups. The former process estimates variance only using
the two groups provided.}

\item{igenes, isamples}{\code{character} vectors with optional subset of
genes which should match \code{rownames(se)}, or samples which should
match \code{colnames(se)}. When all samples from a given sample group are
removed, all associated contrasts are also removed.}

\item{use_voom}{\code{logical} indicating whether to apply \code{limma::voom()}
to define a weight matrix, which is used in the subsequent
analysis steps.}

\item{weights}{optional \code{numeric} matrix used in the limma analysis
workflow, for example the output from \code{limma::voom()}.}

\item{robust}{\code{logical} passed to \code{limma::eBayes()} to indicate
whether to apply robust estimate of variance, for example calling
\code{limma::squeezeVars()}.}

\item{apply_probe_replicates}{\code{logical} indicating whether to combine
probe replicates to one summary result per probe, only when there
are multiple rows that represent the same probe (technical replicates.)
When \code{apply_probe_replicates=TRUE} it applies the limma workflow
with \code{limma::duplicateCorrelation()}, which is applied to every subset
of probes with equivalent number of replicates. The set of probes
which are not replicated is analyzed in its own batch, then all
results are combined in the output. When \code{apply_probe_replicates=TRUE}
the \code{probe_colname} argument should represent a column in
\code{rowData(se)} that contains a repeated identifier that represents
the technical replicates.}
}
\value{
\code{list} whose contents depend upon arguments:
\code{return_df=TRUE} returns one large \code{data.frame} with all contrasts
and stat hit flags applied; \code{return_dfs=TRUE} returns a \code{list} of
\code{data.frames} named by contrast name, each include a summary
stat table, with flags indicating stat hit and direction, the
group mean, and max group mean values; \code{return_array=TRUE} will
return an \code{array} with three dimensions: \code{assay_name} (the name
of the assays analyzed from the input \code{se} object), \code{contrast_name}
with each contrast, and \code{hit_rule} with the stat hit rules;
\code{return_lmfit=TRUE} will return a \code{list} of \code{lmFit} as returned
by \code{limma::lmFit()} and \code{limma::fitContrasts()}.
}
\description{
Run contrasts on SummarizedExperiment data
}
\details{
This function is not yet functional, it is being migrated
from a private set of R functions. Until all required functions
are migrated, this function is not usable.

This function is intended to combine several analysis steps
to test statistical contrasts on data stored in
a \code{SummarizedExperiment} object. Because \code{SummarizedExperiment}
can store multiple assay data matrix entries, for example
with data normalized by different methods, this function
can also test each assay data and assembles results together
for each comparison.
\itemize{
\item Optionally applies \code{limma-voom} workflow.
\item Applies statistical thresholds:
\itemize{
\item P-value, adjusted P-value
\item fold change
\item max group mean (highest mean of groups involved in each contrast)
}
\item Optionally applies a series of different stat hit thresholds
\item Optionally applies stat hits thresholds specific to two-way
contrasts
\item Optionally combines results by gene, using \code{collapse_by_gene()}
\item Assembles a hit array, vectors named by gene or probe, whose
values indicate the sign of the direction of statistical hits
\item Optionally applies recommended limma workflow for replicate
microarray probe spots, for example \code{limma::duplicateCorrelation()}
which returns one summary value per replicated probe
\item Optionally applies a signal floor, replacing values at or below
the floor with the replacement (usually \code{NA} or the floor).
\item Optionally replaces \code{NA} values with an assigned value,
usually \code{0} or the same floor value above.
}
}
