% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{set_xlsx_colwidths}
\alias{set_xlsx_colwidths}
\title{Set column widths in Xlsx files}
\usage{
set_xlsx_colwidths(
  xlsxFile,
  sheet = 1,
  cols = seq_along(widths),
  widths = 11,
  ...
)
}
\description{
Set column widths in Xlsx files
}
\details{
This function is a light wrapper to perform these steps
from the very useful \code{openxlsx} R package:
\itemize{
\item \code{openxlsx::loadWorkbook()}
\item \code{openxlsx::setColWidths()}
\item \code{openxlsx::saveWorkbook()}
}
}
\seealso{
Other jam export functions: 
\code{\link{applyXlsxConditionalFormatByColumn}()},
\code{\link{set_xlsx_rowheights}()}
}
\concept{jam export functions}
