% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{set_xlsx_rowheights}
\alias{set_xlsx_rowheights}
\title{Set row heights in Xlsx files}
\usage{
set_xlsx_rowheights(
  xlsxFile,
  sheet = 1,
  rows = seq_along(heights) + 1,
  heights = 17,
  ...
)
}
\description{
This function is a light wrapper to perform these steps
from the very useful \code{openxlsx} R package:
}
\details{
\itemize{
\item \code{openxlsx::loadWorkbook()}
\item \code{openxlsx::setRowHeights()}
\item \code{openxlsx::saveWorkbook()}
}
}
\seealso{
Other jam export functions: 
\code{\link{applyXlsxConditionalFormatByColumn}()},
\code{\link{set_xlsx_colwidths}()}
}
\concept{jam export functions}
