% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-word-wrap.R
\name{apply_word_wrap}
\alias{apply_word_wrap}
\title{Apply word-wrap to text, recognizing whitespace and some punctuation}
\usage{
apply_word_wrap(x, width = 15, sep = "\\n", ...)
}
\arguments{
\item{x}{\code{character} string or vector.}

\item{width}{\code{numeric} desired maximum character width}

\item{sep}{\code{character} string used as a separator between wrapped
subsets, default \code{"\\n"} wraps to new lines.
When \code{sep=NULL} the wrapped subsets are not combined, instead
are returned as a \code{list} of \code{character} vectors.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{character} vector by default, however when \code{sep=NULL}
it returns a \code{list} of \code{character} vectors.
}
\description{
Apply word-wrap to text, recognizing whitespace and some punctuation
}
\examples{
x <- c("logFC dH1A_VEH_ATAC-UL3_VEH_ATAC",
   "logFC dH1A_VEH_ATAC_UL3_VEH_ATAC",
   "one-two-three-four-five-six-seven")
cat(apply_word_wrap(x), sep="\n--\n")

apply_word_wrap(x, sep=NULL)

}
\concept{jam utility}
