% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-design2colors.R
\name{cardinality}
\alias{cardinality}
\title{Determine cardinality between two vectors}
\usage{
cardinality(x, y = NULL, verbose = FALSE, ...)
}
\description{
Determine cardinality between two vectors
}
\examples{
a <- letters[c(1, 1, 2, 2, 2, 3)];
b <- LETTERS[c(1, 2, 3, 4, 5, 6)];
d <- LETTERS[c(1, 2, 2, 1, 1, 1)];

cardinality(a, b)
cardinality(b, a)

ab <- data.frame(a, b)
ad <- data.frame(a, d)
cardinality(ab, ad)

abt <- tibble::tibble(a, b);
adt <- tibble::tibble(a, d);
cardinality(abt, adt)

abm <- as.matrix(abt);
adm <- as.matrix(adt);
cardinality(abm, adm)

cardinality(d, adm, verbose=TRUE)

cardinality(d, adm, verbose=2)

}
