% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-metabolomics-niehs.R
\name{convert_imputed_assays_to_na}
\alias{convert_imputed_assays_to_na}
\title{Convert imputed metabolomics assay data to NA}
\usage{
convert_imputed_assays_to_na(
  se,
  measurement_assay_names = "area",
  impute_assay_name = "gap_fill_status",
  suffix = "_noimpute",
  impute_flags = c(8, 32),
  ...
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object, or \code{list} of
\code{SummarizedExperiment} objects.
When a \code{list} is provided, each entry is iterated and processed,
and a \code{list} is returned.}

\item{measurement_assay_names}{\code{character} vector with one or more
\code{SummarizedExperiment::assayNames(se)} for which the imputed
data will be converted to \code{NA} values.}

\item{impute_assay_name}{\code{character} string with one
\code{SummarizedExperiment::assayNames(se)} entry that contains the
impute flags. Only the first matching assay name is used.}

\item{suffix}{\code{character} suffix appended to each value in
\code{measurement_assay_names} to create a new assay name for the
resulting data matrix.

When filtering by different impute criteria, it may be helpful
to include the codes in the suffix, for example filtering by only
\code{impute_flags=8} may suggest \code{suffix="_noimpute8"}.}

\item{impute_flags}{\code{numeric} vector of flags indicating values in
\code{impute_assay_name} assay data which represent imputed data.

The defaults:
\itemize{
\item 8: filled by trace area (imputed)
\item 32: filled by spectrum noise (imputed)
}

Other values:
\itemize{
\item 0: original ion used
\item 64: filled by matching ion
\item 128: filled by re-detected peak
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{SummarizedExperiment} or when input \code{se} is a \code{list}
a \code{list} of \code{SummarizedExperiment} objects is returned.
}
\description{
Convert imputed metabolomics assay data to NA
}
\details{
This function takes \code{SummarizedExperiment} data as input, which
must contain at least two \code{SummarizedExperiment::assayNames()}
entries:
\enumerate{
\item \code{numeric} assay measurements
\item \code{numeric} impute flag data
}

The coordinates of cells in the impute flag data
which match \code{impute_flags} are used to convert data in assay measurements
to \code{NA}.
}
\seealso{
Other jam utility functions: 
\code{\link{color_complement}()},
\code{\link{curate_se_colData}()},
\code{\link{curate_to_df_by_pattern}()},
\code{\link{get_numeric_transform}()},
\code{\link{merge_proteomics_se}()},
\code{\link{rowNormScale}()}
}
\concept{jam utility functions}
