% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-proteomics.R
\name{convert_PD_df_to_SE}
\alias{convert_PD_df_to_SE}
\title{Internal function to convert Proteomics Discoverer data.frame to SummarizedExperiment}
\usage{
convert_PD_df_to_SE(
  protein_df,
  ann_lib = c("org.Hs.eg.db"),
  curation_txt = NULL,
  ptm_colname = "Modifications",
  type = c("protein", "peptide"),
  remove_duplicate_peptides = TRUE,
  accession_from = NULL,
  accession_to = NULL,
  xref_df = NULL,
  verbose = FALSE,
  ...
)
}
\description{
Internal function to convert Proteomics Discoverer data.frame to SummarizedExperiment
}
\details{
This function is intended to be called by \code{import_proteomics_PD()} after
the Excel data is split into protein and peptide \code{data.frame} components.
}
