% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-deeptools.R
\name{deepTools_matrix2nmat}
\alias{deepTools_matrix2nmat}
\title{Import deepTools coverage matrix to normalizedMatrix}
\usage{
deepTools_matrix2nmat(
  x = NULL,
  filename = NULL,
  signal_name = NULL,
  target_name = NULL,
  background = 0,
  smooth = FALSE,
  target_is_single_point = FALSE,
  signal_is_categorical = FALSE,
  mat_grep = "[-0-9]+:[-0-9]+",
  upstream_grep = "^[-]",
  downstream_grep = "^[^-]",
  target_grep = "^0$",
  verbose = FALSE,
  ...
)
}
\description{
Import deepTools coverage matrix to normalizedMatrix
}
\details{
This function imports deepTools matrix data to produce
a \code{normalizedMatrix} object as defined in EnrichedHeatmap.

The initial use case is to import binned data across a fixed range,
relative to some central feature "target", which is defined
using upstream distance, downstream distance, and bins defined
with fixed width.

Minor modifications were made previously to accommodate scaled
coverage bins. More development will be done to handle the
column headers properly in future.
\subsection{Supported features}{
\itemize{
\item Recognizes upstream, downstream, bin size
\item Recognizes sample labels
\itemize{
\item multiple samples are split into a \code{list} of matrix objects
}
\item Recognizes row groups
\itemize{
\item Row groups are stored in \code{attr(x, "anno_df")} of the returned object
}
\item Duplicated rownames are adjusted using \code{jamba::makeNames()}
}
}
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_lipotype_csv}()},
\code{\link{import_metabolomics_niehs}()},
\code{\link{import_nanostring_csv}()},
\code{\link{import_nanostring_rcc}()},
\code{\link{import_nanostring_rlf}()},
\code{\link{import_omics_data}()},
\code{\link{import_proteomics_PD}()},
\code{\link{import_proteomics_mascot}()},
\code{\link{import_salmon_quant}()},
\code{\link{process_metab_compounds_file}()}
}
\concept{jam import functions}
