% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-design2colors.R
\name{design2colors}
\alias{design2colors}
\title{Convert experiment design into categorical colors}
\usage{
design2colors(
  x,
  group_colnames = NULL,
  lightness_colnames = NULL,
  class_colnames = NULL,
  ignore_colnames = NULL,
  color_list = NULL,
  color_sub = NULL,
  color_sub_max = NULL,
  preset = "dichromat2",
  phase = 1,
  rotate_phase = -1,
  class_pad = 2,
  end_hue_pad = 0,
  hue_offset = 0,
  desat = c(0, 0.4),
  dex = c(2, 5),
  Crange = NULL,
  Lrange = NULL,
  na_color = "grey90",
  use_naValue = "<NA>",
  shuffle_colors = FALSE,
  force_consistent_colors = TRUE,
  use_add_colors = FALSE,
  plot_type = c("table", "list", "none"),
  return_type = c("list", "df", "vector"),
  verbose = FALSE,
  debug = c("none", "cardinality"),
  ...
)
}
\arguments{
\item{x}{\code{data.frame} with columns to be colorized,
\code{DataFrame} from Bioconductor \code{S4Vectors} package,
\code{tbl_df} from the \code{tibble} package, or \code{matrix}. In all cases
the data is coerced to \code{data.frame} without changing colnames,
and without imposing factor values per column, unless factors
were already encoded.}

\item{group_colnames}{\code{character} or \code{intger} vector indicating
which \code{colnames(x)} to use, in order, for group color assignment.}

\item{lightness_colnames}{\code{character} or \code{intger} vector indicating
which \code{colnames(x)} to use, in order, for group lightness gradient.}

\item{class_colnames}{\code{character} or \code{intger} vector indicating
higher-level grouping of \code{group_colnames}}

\item{color_list}{\code{list} default NULL, with names in \code{colnames(x)}
with optional color assignments by column. This argument represents
input recognized by \code{heatmap_se()} argument \code{sample_color_list}.
Each elementmay be one of the following:
\itemize{
\item \code{character} vector of colors, with names matching values in
the column.
\item \code{character} with single color, to be applied as a gradient to
column values.
\item \code{function} which takes column values, typically numeric, and
returns one color for each input value.
}}

\item{color_sub}{one of the following:
\itemize{
\item \code{character} vector of R colors, where \code{names(color_sub)}
assign each color to a character string. It is intended to allow
specific color assignments upfront.
Also \code{names(color_sub)} may match \code{colnames(x)} in which case
the color is assigned to that color using a color gradient across
the unique character values in that column. Values are assigned in
order of their appearance in \code{x} unless the column is a \code{factor},
in which case colors are assigned to \code{levels} in order.
\item \code{list} named by \code{colnames(x)} to assign colors to each column.
In this case, values may be: \code{character} vector with colors named
by unique value in the column; \code{function} to apply to values in the
column, assumed to be \code{numeric} values; or single \code{character} color
which is used to create a color gradient as described above.
}}

\item{color_sub_max}{\code{numeric} optional value used to define a fixed
upper limit to a color gradient when a color is applied to
a \code{numeric} column.
\itemize{
\item When one value is defined for \code{color_sub_max} it is used for all
\code{numeric} columns uniformly.
\item When multiple values are defined for \code{color_sub_max}, then
\code{names(color_sub_max)} are used to associate to the appropriate
column matching with \code{colnames(x)}.
}}

\item{preset}{\code{character} string passed to \code{colorjam::h2hwOptions()},
which defines the hues around a color wheel, used when selecting
categorical colors. Some shortcuts:
\itemize{
\item \code{"dichromat"}: (default) uses color-blindness friendly color wheel,
minimizing effects of three types of color blindness mainly by
removing large chunks of the green color space.
\item \code{"ryb"}: red-yellow-blue color wheel, which emphasizes the yellow
part of the wheel as a major color, as opposed to computer
monitor default that represents the red-green-blue color components.
\item \code{"ryb2"}: red-yellow-blue color wheel, version 2, adjusted
to reduce effects of greenish-yellow for aesthetics.
\item \code{"rgb"}: default red-green-blue color wheel used by computer
monitors to mimic the components of human vision.
}}

\item{phase, rotate_phase}{\code{integer} value, \code{phase} is passed to
\code{colorjam::rainbowJam()} to define the light/dark pattern phasing,
which has 6 positions, and negative values reverse the order.
Categorical colors are assigned to the class/group combinations,
after which \code{phase + rotate_phase} is used for categorical colors
for any remaining values.}

\item{class_pad}{\code{integer} zero or greater, indicating the number
of empty hues to insert as a spacer between hues when the class
changes in a sequence of class/group values. Higher values will
ensure the hues in each class are more distinct from each other
across class, and more similar to each other within class.}

\item{end_hue_pad}{\code{integer} used to pad hues at the end of a
color wheel sequence, typically useful to ensure the last color
is not similar to the first color.}

\item{desat}{\code{numeric} vector extended to length=2, used to desaturate
class/group colors, then remaining colors, in order. The intended
effect is to have class/group colors visibly more colorful than
remaining colors assigned to other factors.}

\item{dex}{\code{numeric} vector passed to \code{jamba::color2gradient()} to
define the darkness expansion factor, where 1 applies a moderate
effect, and higher values apply more dramatic light-to-dark
effect. When \code{dex} has length=2, the second value is used only
for columns where colors are assigned by \code{colnames(x)}
using \code{color_sub}.}

\item{Crange, Lrange}{\code{numeric} ranges passed to \code{colorjam::rainbowJam()}
to define slightly less saturated colors than the default rainbow
palette.}

\item{na_color}{\code{character} string, default "grey90", R color assigned
to \code{NA} values.}

\item{use_naValue}{\code{character} string, default \code{"<NA>"}, used to assign
a name to the color defined by \code{na_color}.}

\item{shuffle_colors}{\code{logical}, default FALSE, whether to shuffle
color assignments for values not already assigned by
class/group/lightness, nor by \code{color_sub}.
The effect is that colors are less likely to be similar for
adjacent column values.}

\item{force_consistent_colors}{\code{logical}, default TRUE, whether to force
color substitutions for the same character string across columns.
Note: This scenario is most
likely to occur when using \code{color_sub} to assign colors to a
specific column in \code{colnames(x)}, and where that column may contain
one or more values already assigned a color earlier in the process.
For example: class/group/lightness defines colors for these columns,
then all columns are checked for cardinality with these color
assignments. Columns not appearing in \code{color_sub} will be colorized
when their cardinality is compatible with group colors, otherwise
specific values may be assigned colors via \code{color_sub}, then all
remaining values are assigned categorical colors. This process defines
colors for each column. The last step reassigns colors consistently
using the first value-color association that appears in the list,
to make sure all assignments are consistent. This last step is subject
of the argument \code{force_consistent_colors}. In either case, the
output \code{color_sub} will only have one color assigned per value.
\itemize{
\item Default \code{TRUE}: When colors are being assigned to column values
\code{color_sub}, if the value had been assigned a color in a previous
column, for example by \code{group_colnames} color assignment, then the
first assignment is applied to all subsequent assignments.
\item Optional \code{FALSE}: Color assignments are re-used where applicable,
except when overridden by \code{color_sub} for a particular column. In
that case, color assignments are maintained for each specific column.
}}

\item{use_add_colors}{\code{logical} whether to use experimental
\code{colorjam::add_colors()} which finds colors with maximum perceptual
difference from the existing colors. It is notably slower than other
approaches, and does not give options to provide specific color hues.}

\item{plot_type}{\code{character} string indicating a type of plot for results:
\itemize{
\item \code{"table"}: plots a color table equal to the input \code{data.frame} where
background cells indicate color assignments.
\item \code{"list"}: plots colors defined for each column in \code{x} using
\code{jamba::showColors()}
\item \code{"none"}: no plot is produced
}}

\item{return_type}{\code{character} string indicating the data format to return:
\itemize{
\item \code{"list"}: a \code{list} of colors, named by \code{colnames(x)}.
\item \code{"df"}: a \code{data.frame} in order of \code{x} with colors assigned to each cell.
\item \code{"vector"}: a \code{character} vector of R colors, named by assigned
factor level.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{debug}{\code{character} string used to enable detailed debugging output.}

\item{...}{additional arguments are passed to downstream functions.}
}
\value{
output depends upon argument \code{return_type}:
\itemize{
\item \code{"list"}: returns a \code{list} of colors defined by \code{colnames(x)},
suitable for use with \code{ComplexHeatmap::HeatmapAnnotation()} for example.
\item \code{"df"}: returns \code{data.frame} of colors with same dimensions as the
input \code{x}. Suitable for use with \code{jamba::imageByColors()} for example.
\item \code{"vector"}: returns \code{character} vector of R colors, whose names represent
values in \code{x}, where the values should be substituted with the color.
Suitable for use with \code{ggplot2::color_manual(values=colors)}.

In all cases, the \code{attributes()} of the returned object also includes
colors in the other two formats: \code{"color_sub"}, \code{"color_df"}, and
\code{"color_list"}.
}
}
\description{
Convert experiment design into categorical colors
}
\details{
The general goal is to assign categorical colors relevant to
the experimental design of an analysis. The basic logic:
\enumerate{
\item Assign categorical colors to experimental groups.
\item Shade these colors light-to-dark based upon secondary factors.
\item For step 1 above, optionally assign similar color hues by class.
}

When there are multiple factors in a design, the general guidance:
\itemize{
\item Define \code{group_colnames} using the first two factors in the design.
\item Define \code{class_colnames} using one of these two factors.
Values in \code{group_colnames} will be assigned rainbow categorical colors,
with extra spacing between each class. Values in one class will be
assigned similar color hues, for example one class may be red/orange,
another class may be blue/purple.
\item Optionally choose another factor to use as \code{lightness_colnames}.
When there are multiple unique values per group, they will be
shaded from light to dark within the group color hue.
}

It is sometimes helpful to create a column for \code{class_colnames},
for example when a series of treatments can be categorized
by the type of treatment (agonists, antagonists, inhibitors,
controls).

Franky, we tend to try a few combinations until the output seems
intuitive. Then we assign specific values from other columns
using \code{color_sub}. Typically for \code{numeric} columns we assign
a color to the colname, and for \code{categorical} colors we assign
colors to values in the column.

Version 0.0.69.900 and higher: When the cardinality of group/class
values is not 1-to-many, either the group/class assignments
are switched in order to create 1-to-many cardinality, or
a combination of the two vectors is used to create the appropriate
cardinality.
\subsection{When no group_colnames or class_colnames are defined}{

By default, the unique rownames are used as if they were groups,
then colors are assigned using the same logic as usual. Any
other column whose values are 1-to-1 match with rownames will
inherit the same colors, otherwise \code{character} and \code{factor}
columns will be assigned categorical colors, and \code{numeric}
columns will be assigned a color gradient.
}

\subsection{Categorical colors}{

At its simplest a set of groups can be assigned categorical colors.
\itemize{
\item colors should be visibly distinct from one another
\item colors should generally be distinct across forms of color-blindness
\item colors should be consistent across plots, figures, tables
}

Finally, colors may be pre-defined using a named vector of colors.
These colors will be propagated to other entries in the table.
}

\subsection{Light-to-dark gradient}{

The light-to-dark gradient is intended for ordered sub-divisions,
for example:
\itemize{
\item across time points in a time series
\item across treatment doses in an ordered series
\item across ordered measurements first-to-last
}
}

\subsection{Group class}{

The group classification is intended to assign color hues
for similar groups:
\itemize{
\item antagonists, agonists, untreated
\item treated, untreated
\item wildtype, mutant form 1, mutant form 2, etc.
}

For example, antagonists may be assigned colors blue-to-purple;
agonists may be assigned colors red-to-orange; with a pronounced
color hue "gap" between antagonists and agonists.
}

\subsection{Additional categorical color assignment}{

Finally, other annotations associated with samples are assigned
categorical colors, visibly distinct from other color assignments.

For entries associated with only one design color, for example "Sample_ID",
"Sample Name", "Lane Number", or "Well Number",
they inherit the design color.

For entries associated with more than one design color, for example
"Batch", "Date", or perhaps "Dose", they will be assigned a unique
color.
\itemize{
\item additional annotations unique to design colors inherit the design colors
\item additional categorical colors should not duplicate existing colors
}
}

\subsection{Future ideas}{
\itemize{
\item Assign "additional factors" to colors based upon \code{class}
\itemize{
\item Currently "additional factors" are only tested by class_group and
class_group_lightness.
\item It could be useful to test versus \code{class} alone (if supplied)
\item Goal would be to assign color hue using the mean color hue in the class.
\item Otherwise the class may be assigned a color inconsistent with the
range of color hues.
}
\item Handle numeric columns by applying color gradient
\itemize{
\item A truly numeric column (not just integer index values) could
use \code{circlize::colorRamp2()} to apply color gradient
}
}
}
}
\examples{
df <- data.frame(
   genotype=rep(c("WT", "GeneAKO", "GeneBKO"), c(4, 8, 8)),
   treatment=rep(rep(c("control", "treated"), each=2), 5),
   class=rep(c("WT", "KO"), c(4, 16)),
   time=c(rep("early", 4),
      rep(rep(c("early", "late"), each=4), 2)))
df$sample_group <- jamba::pasteByRow(df[,c("genotype", "treatment", "time")])
df$sample_name <- jamba::makeNames(df$sample_group);
df$age <- sample(40:80, size=nrow(df));
df

dfc <- design2colors(df,
   group_colnames="genotype",
   lightness_colnames="treatment",
   class_colnames="class",
   color_sub=c(age="dodgerblue"))

# same as above except use add_colors()
dfca <- design2colors(df,
   group_colnames="genotype",
   use_add_colors=TRUE,
   lightness_colnames="treatment",
   class_colnames="class",
   color_sub=c(age="dodgerblue"))

# same as above except assign colors to columns and some values
dfc <- design2colors(df,
   group_colnames="sample_group",
   lightness_colnames="treatment",
   class_colnames="genotype",
   class_pad=5,
   preset="dichromat2",
   color_sub=c(KO="darkorchid3",
      treatment="navy",
      time="dodgerblue"))

# same as above except assign specific group colors
dfc <- design2colors(df,
   group_colnames="genotype",
   lightness_colnames="treatment",
   class_colnames="class",
   preset="dichromat2",
   color_sub=c(
      WT="gold",
      KO="purple3",
      age="firebrick",
      GeneAKO="firebrick3",
      GeneBKO="dodgerblue",
      treatment="navy",
      time="darkorchid4"))

dfc2 <- design2colors(df,
   group_colnames="genotype",
   lightness_colnames=c("time", "treatment"),
   class_colnames="class",
   preset="dichromat2")

dfc3 <- design2colors(df,
   group_colnames=c("genotype"),
   lightness_colnames=c("time", "treatment"),
   class_colnames="genotype",
   preset="dichromat2")

df1 <- df;
df2 <- subset(df, time \%in\% "early");
df12 <- rbind(df1, df2);
dfc12 <- design2colors(df12,
   group_colnames="genotype",
   lightness_colnames=c("time", "treatment"),
   class_colnames="class",
   preset="dichromat",
   use_add_colors=TRUE,
   color_sub=c(
      treatment="steelblue",
      time="dodgerblue"
   ))


}
\seealso{
Other jam color functions:
\code{\link[=assign_numeric_colors]{assign_numeric_colors()}},
\code{\link[=df_to_numcolors]{df_to_numcolors()}},
\code{\link[=mean_hue]{mean_hue()}},
\code{\link[=print_color_list]{print_color_list()}},
\code{\link[=quick_complement_color]{quick_complement_color()}}
}
\concept{jam color functions}
