% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-design2colors.R
\name{design2colors}
\alias{design2colors}
\title{Convert experiment design into categorical colors}
\usage{
design2colors(
  x,
  group_colnames = NULL,
  lightness_colnames = NULL,
  class_colnames = NULL,
  preset = "dichromat",
  phase = 1,
  rotate_phase = -1,
  class_pad = 1,
  end_hue_pad = 2,
  desat = 0,
  plot_type = c("table", "list", "none"),
  return_type = c("list", "df", "vector"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} with columns to be colorized}

\item{group_colnames}{\code{character} or \code{intger} vector indicating
which \code{colnames(x)} to use, in order, for group color assignment.}

\item{lightness_colnames}{\code{character} or \code{intger} vector indicating
which \code{colnames(x)} to use, in order, for group lightness gradient.}

\item{class_colnames}{\code{character} or \code{intger} vector indicating
higher-level grouping of \code{group_colnames}}

\item{...}{additional arguments are passed to downstream functions.}
}
\description{
Convert experiment design into categorical colors
}
\examples{
df <- data.frame(
   genotype=rep(c("WT", "GeneAKO", "GeneBKO"), c(4, 8, 8)),
   treatment=rep(rep(c("control", "treated"), each=2), 5),
   class=rep(c("WT", "KO"), c(4, 16)),
   time=c(rep("early", 4),
      rep(rep(c("early", "late"), each=4), 2)))
df$sample_group <- jamba::pasteByRow(df[,c("genotype", "treatment", "time")])
df$sample_name <- jamba::makeNames(df$sample_group);
df

dfc <- design2colors(df,
   group_colnames="genotype",
   lightness_colnames="treatment",
   class_colnames="class",
   preset="rgb")

dfc2 <- design2colors(df,
   group_colnames="genotype",
   lightness_colnames=c("time", "treatment"),
   class_colnames="class",
   preset="dichromat")

dfc3 <- design2colors(df,
   group_colnames=c("genotype"),
   lightness_colnames=c("time", "treatment"),
   class_colnames="genotype",
   rotate_phase=TRUE,
   preset="dichromat")

}
