% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-design2layout.R
\name{design2layout}
\alias{design2layout}
\title{Convert experiment design into reasonable plot panel layout}
\usage{
design2layout(SE, group_colname = NULL, run_colname = NULL, ...)
}
\value{
\code{list} with elements:
\itemize{
\item \code{nrow}: number of panel rows
\item \code{ncol}: number of panel columns
\item \code{blank_pos}: integer vector for blank panels, if any
\item \code{Group_ctm}: summary info about the panel layout
\item \code{isamples}: character vector indicating order to plot
}
}
\description{
Convert experiment design into reasonable plot panel layout
}
\details{
This function is in development, and is intended for a specific
scenario where an experiment design with optional batch/run
can be represented in a sensible layout, with appropriate
number of panel columns, and number of panel rows.
}
\seealso{
Other jam utility functions: 
\code{\link{cardinality}()},
\code{\link{color_complement}()},
\code{\link{convert_PD_df_to_SE}()},
\code{\link{convert_imputed_assays_to_na}()},
\code{\link{curate_se_colData}()},
\code{\link{curate_to_df_by_pattern}()},
\code{\link{get_numeric_transform}()},
\code{\link{handle_df_args}()},
\code{\link{merge_proteomics_se}()},
\code{\link{rmd_tab_iterator}()},
\code{\link{rowNormScale}()}
}
\concept{jam utility functions}
