% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-design2colors.R
\name{df_to_numcolors}
\alias{df_to_numcolors}
\title{Convert data.frame to numeric color substitution}
\usage{
df_to_numcolors(df, colramp = c("Reds"), colramp_divergent = c("RdBu_r"), ...)
}
\description{
Convert data.frame to numeric color substitution
}
