% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-neuro.R
\name{frequency_matrix2nmat}
\alias{frequency_matrix2nmat}
\title{Convert frequency matrix to normalizedMatrix format}
\usage{
frequency_matrix2nmat(
  mat,
  target_frequency = NULL,
  target_name = "target",
  signal_name = "Frequency",
  do_scale = TRUE,
  scale_from = 0,
  scale_to = 1,
  scale_frequency = NULL,
  apply_floor = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mat}{numeric matrix with frequency represented as
columns, and whose numeric frequency is stored in
\code{colnames(mat)} as character values.}

\item{target_frequency}{numeric vector representing a
range of frequencies to be considered the "target",
and thus highlighted by \code{EnrichedHeatmap::EnrichedHeatmap()}.
By default the target range is used to order rows from
high to low signal, unless the order is specified
otherwise. When \code{target_frequency} is \code{NULL}, there
is no target indicated in the normalizedMatrix.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{normalizedMatrix} object, a subclass of \code{matrix},
as defined in \code{EnrichedHeatmap}.
}
\description{
Convert frequency matrix to normalizedMatrix format
}
\details{
This function takes input data with frequency represented
as columns, and observations as rows. It will optionally
scale each row to have fixed minimum-to-maximum value
range, given a range of frequencies to use for scaling.
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{import_nanostring_rcc}()}
}
\concept{jam import functions}
