% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import.R
\name{get_nmat_ceiling}
\alias{get_nmat_ceiling}
\title{Helper function to calculate signal ceiling of numeric matrix}
\usage{
get_nmat_ceiling(imat, iceiling = NULL, verbose = TRUE, ...)
}
\description{
Helper function to calculate signal ceiling of numeric matrix
}
\details{
This function is called by \code{nmatlist2heatmaps()} and is not
intended to be called directly.

It takes a \code{normalizedMatrix} or \code{numeric} matrix object, and
a ceiling value \code{iceiling} and determines an appropriate numeric
ceiling with the following rules:
\itemize{
\item if \code{iceiling} is \code{NULL}, it returns the highest absolute value in \code{imat}
\item if \code{iceiling > 0} and \code{iceiling <= 1}, it calculates
\code{quantile(abs(imat), probs=iceiline)}, excluding values of zero
\item otherwise \code{iceiling} is used as a fixed numerical ceiling
}

In all cases, \code{iceiling} is rounded to 3 digits with \code{round(iceiling, digits=3)}

Also in all cases, \code{na.rm=TRUE} is used, to prevent returning \code{NA}.
}
\seealso{
Other jam coverage heatmap functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{nmathm_row_order}()},
\code{\link{nmatlist2heatmaps}()},
\code{\link{validate_heatmap_params}()},
\code{\link{zoom_nmatlist}()},
\code{\link{zoom_nmat}()}
}
\concept{jam coverage heatmap functions}
