% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import.R
\name{get_nmat_ceiling}
\alias{get_nmat_ceiling}
\title{Helper function to calculate signal ceiling of numeric matrix}
\usage{
get_nmat_ceiling(imat, iceiling = NULL, verbose = TRUE, ...)
}
\description{
Helper function to calculate signal ceiling of numeric matrix
}
\details{
This function is called by \code{nmatlist2heatmaps()} and is not
intended to be called directly.
}
\seealso{
Other jam utility functions: 
\code{\link{get_numeric_transform}()},
\code{\link{rowNormScale}()}
}
\concept{jam utility functions}
