% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import.R
\name{get_numeric_transform}
\alias{get_numeric_transform}
\title{Get appropriate numeric transformation function}
\usage{
get_numeric_transform(transform, ...)
}
\arguments{
\item{transform}{\code{character} string or \code{function}. If
\code{length(transform) > 1} then \code{get_numeric_transform()}
is called for each value in \code{transform}. Valid names:
\code{"log2signed"} applies \code{jamba::log2signed()} which applies
\code{log2(1+x)} transform to the absolute value, then multiplies
by the original \code{sign(x)};
\code{"exp2signed"} applies the inverse of \code{"log2signed"}, which
exponentiates numeric values which were previously transformed
with \code{log2(1+x)};
\code{"sqrt"} applies square root transform;
\code{"cubert"} applies cube root \code{x^(1/3)};
\code{"qrt"} applies fourth root \code{x^(1/4)};
\code{"frt"} applies fourth root \code{x^(1/5)};
\code{"square"} applies \code{x^2} to absolute value, multiplied by the \code{sign(x)};
\code{"cube"} applies \code{x^3};
When there are negative numeric
values, the transformation is applied to absolute value,
then multiplied by the original sign. Therefore, the
transformation is applied to adjust the magnitude of
the values.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{function} or \code{NULL} when no matching function is
found, or \code{list} is returned when the input \code{transform}
has multiple values.
}
\description{
Get appropriate numeric transformation function
}
\details{
This function recognizes numeric transformation functions by
name, or searches the attached R package environments for
a matching function by name.
}
\examples{
get_numeric_transform("log2signed")

}
