% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{get_salmon_root}
\alias{get_salmon_root}
\title{Get Salmon root directory}
\usage{
get_salmon_root(file)
}
\arguments{
\item{file}{character path to a Salmon output file, or a vector
of files.}
}
\value{
file path to the Salmon root directory, or \code{NA} if
the Salmon root directory could not be found.
}
\description{
Get Salmon root directory
}
\details{
This function uses the \code{rprojroot} package to find the
root output directory of a Salmon quant analysis, looking
for the file \code{"cmd_info.json"}. If this file is not found,
this function returns \code{NULL}.
}
\examples{
cmdinfopath <- system.file("data", "salmonOut", "cmd_info.json", package="platjam");
if (nchar(cmdinfopath) > 0) {
   get_salmon_root(cmdinfopath);
}

}
\seealso{
Other jam nextgen sequence functions: 
\code{\link{get_salmon_meta}()}
}
\concept{jam nextgen sequence functions}
