% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ucsctemplates.R
\name{get_track_defaults}
\alias{get_track_defaults}
\title{Get UCSC track default values and track templates}
\usage{
get_track_defaults(env = new.env())
}
\arguments{
\item{env}{\code{environment} in which to store the default values
and track templates.}
}
\value{
\code{environment} which contains the default values, and template
values.
}
\description{
Get UCSC track default values and track templates
}
\details{
This function defines default values for overlay and composite
track types. It also defines three templates each for overlay
and composite:
\itemize{
\item overlay_header - equivalent to superTrack
\item overlay_parent - equivalent to one set of overlay tracks
\item overlay_track - each overlay track
\item composite_header - equivalent to one composite track
\item composite_parent - equivalent to a composite view within a composite track
\item composite_track - individual track within the composite track.
}
}
\seealso{
Other jam ucsc browser functions: 
\code{\link{assign_track_defaults}()},
\code{\link{make_ucsc_trackname}()},
\code{\link{parse_ucsc_gokey}()}
}
\concept{jam ucsc browser functions}
