% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-geomx.R
\name{import_geomx_samplesheet}
\alias{import_geomx_samplesheet}
\title{Import GeoMx SampleSheet.csv data}
\usage{
import_geomx_samplesheet(
  x = "SampleSheet.csv",
  return_type = c("dflist", "list", "indices", "filenames"),
  do_revcomp = TRUE,
  demux_sheetnumber = "S1",
  demux_prefix = "BK-Gq-1_hdist1_",
  ...
)
}
\arguments{
\item{x}{\code{character} path to SampleSheet.csv formatted GeoMx file.}

\item{return_type}{\code{character} string to define the return type:
\itemize{
\item \code{"dflist"} - \code{list} of \code{data.frame} for each heading (default)
\item \code{"list"} - \code{list} of lines as-is.
\item \code{"indices"} - \code{character} vector of expected indices, in format
\code{"Index+Index2"}.
\item \code{"filenames"} - \code{data.frame} describing the expected output filenames
after running \code{demuxbyname.sh} (BBTools), and the expected
GeoMx filename using the Sample_ID, suitable to rename one file
to the other.
}}

\item{do_revcomp}{\code{logical} indicating whether to use reverse complement
for Index2, default=TRUE.}

\item{demux_sheetnumber}{\code{character} used when \code{return_type="filenames"}
in formulating the correct GeoMx input filename for the NGS pipeline.
This value therefore affects the filename after renaming the demux
file, to become the NGS input filename.}

\item{demux_prefix}{\code{character} prefix assigned when running
\code{demuxbyname.sh} (BBTools). It is usually in the form
\code{"project_hist1"} where \code{"hist1"} refers to the Hamming distance
threshold \code{"hdist"} used with \code{demuxbyname.sh}.
This prefix therefore is used to match the filename produced by
\code{demuxbyname.sh}, which is then renamed to the NGS filename.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} or \code{character} vector, consistent with \code{return_type}.
}
\description{
Import GeoMx SampleSheet.csv data
}
\details{
This function can import \code{"SampleSheet.csv"} and \code{"GNP_config.ini"}
files, which are characterized as follows:
\itemize{
\item Each subset of data is preceded by a header line: \verb{[header_name]}
\item Data following this line is comma-delimited, or delimited with
\code{" = "}, both of which are treated as equivalent.
\item There is a blank line between the subset of data and the next header.
}

Some rules with \code{return_type="dflist"}:
\itemize{
\item Delimiters are recognized as \code{" = "} or \code{","}, but because the
import process calls \code{data.table::fread()} it will probably also
accept tab-delimited data.
\item When the first row following the header appears to have column names,
they are used as-is as column names.
\item The following criteria cause the first row NOT to be used as
column header:
\itemize{
\item The first entry does not contain \code{"Sample_ID"}, and any of:
\item Any value following a comma begins with a number, or
\item Any value following a comma is \code{"true"} or \code{"false"}, or
\item Any value following a comma is purely DNA sequence \code{"[ATGC]+"}, or
\item There is no \code{","} delimiter, or
\item There is only one value in the subset of data.
}
}

Therefore, when the first entry begins with \code{"Sample_ID"} the
first entry is used as column header.
\itemize{
\item When the first entry is not used as column headers, the heading name
itself is used as the first column name, followed by \code{V} concatenated
to the integer column number, for example: \verb{"Sequencing", "V2", "V3", "V4"}
}
\subsection{Todo:}{
\itemize{
\item Add \code{return_type} option to create commands to rename demux output
files to the expected GeoMx Sample_ID format.
}
}
}
\examples{
samplefile <- system.file("data", "SampleSheet.csv", package="platjam")
samplelist <- import_geomx_samplesheet(samplefile)
lengths(samplelist)

}
\seealso{
Other jam GeoMx functions: 
\code{\link{revcomp}()}
}
\concept{jam GeoMx functions}
