% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-lipotype.R
\name{import_lipotype_csv}
\alias{import_lipotype_csv}
\title{Import LipoType lipidomics data from CSV file}
\usage{
import_lipotype_csv(
  csv,
  type = "raw",
  curation_txt = NULL,
  num_header_rows = 3,
  num_anno_columns = 7,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{csv}{\code{character} path to a comma-delimited file (csv) as
provided in a LipoType summary report.}

\item{curation_txt}{\code{data.frame} whose first column should match the
column headers found in the LipoType measurement data.
If \code{curation_txt} is not supplied, then values will be split into
columns by \verb{_} underscore or \code{" "} whitespace characters.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{SummarizedExperiment} where:
\itemize{
\item \code{rowData} represents lipid measurement annotations
\item \code{colData} represents sample annotations, optionally including
annotations via a \code{data.frame} supplied as \code{curation_txt}.
}
}
\description{
Import LipoType lipidomics data from CSV file
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_nanostring_rcc}()},
\code{\link{import_proteomics_PD}()},
\code{\link{import_salmon_quant}()}
}
\concept{jam import functions}
