% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-nanostring-csv.R
\name{import_nanostring_csv}
\alias{import_nanostring_csv}
\title{Import Nanostring csv file}
\usage{
import_nanostring_csv(
  csv,
  probe_colname = "Probe_ID",
  probe_anno_file = NULL,
  assay_name = NULL,
  hk_count = 10,
  curation_txt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{csv}{\code{character} path to CSV file, suitable for import
by \code{data.table::fread()}, so most delimiters will be recognized.}

\item{probe_colname}{\code{character} string used to define the probe
identifier in \code{rowData()}, derived from the \code{colnames()} of the
import data. Sometimes the values represent \code{"Probe_ID"}, and
sometimes they may represent \code{"Gene"}, \code{"miRNA"} or another
measurement.}

\item{probe_anno_file}{\code{character} path to one or more files with
additional annotation for each measurement, with identifiers that
match \code{probe_colname} above.}

\item{assay_name}{\code{character} string of optional assay name used
in the \code{SummarizedExperiment} object created. When \code{NULL} the
default is \code{"raw"} unless the file contains substring \code{"norm"}
in which case it uses \code{"norm"}.}

\item{hk_count}{\code{integer} number of housekeeper genes to use when
\code{"control_type"} is not defined for the imported data.
In this case, the last \code{hk_count} genes in the data are
assumed to be housekeeper genes, by typical convention of
NanoString codeset design.}

\item{curation_txt}{either \code{data.frame} or \code{character} file path
to tab- or comma-delimited file. The first column should match the
column headers after importing data, \code{colData(se)}.
Subsequent columns contain associated sample annotations.
For Nanostring data, the Nanostring sample annotations will already
be associated with the \code{colData(se)}, and \code{colnames(curation_df)}
will overwrite any that already exist.
Pro tip: The first column in \code{curation_txt} should contain \code{'.'}
instead of punctuation/whitespace, to improve pattern matching
filenames where the punctuation characters may have been modified
during processing.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{SummarizedExperiment}
}
\description{
Import Nanostring csv file
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_lipotype_csv}()},
\code{\link{import_nanostring_rcc}()},
\code{\link{import_proteomics_PD}()},
\code{\link{import_salmon_quant}()}
}
\concept{jam import functions}
