% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-omics-data.R
\name{import_omics_data}
\alias{import_omics_data}
\title{Import generic omics tsv or csv file}
\usage{
import_omics_data(
  x,
  assay_name = "data",
  row_identifier = NULL,
  row_annotation_columns = NULL,
  curation_txt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{one of the following input types:
\itemize{
\item \code{character} path to a data file suitable for import
by \code{data.table::fread()}, where most delimiters will be recognized
automatically.
\item \code{data.frame} representing the equivalent data as from a file.
\item \code{matrix} representing only the measured values, with rownames
and colnames that represent respective identifiers.
}}

\item{assay_name}{\code{character} string which will define the assay name
where data is stored in the resulting \code{SummarizedExperiment} object.}

\item{row_identifier}{defines the column, columns, or rownames, to be
used as row identifiers. These values become \code{rownames(se)} for
the output object.
One of the following:
\itemize{
\item \code{NULL} (default) which auto-detects an appropriate row identifier:
\itemize{
\item If the first column is numeric, it uses rownames.
\item Otherwise the first column is used. If there are duplicated values,
they are made unique with \code{jamba::makeNames()}.
}
\item \code{character} string of colname in the data \code{x} imported.
\item \code{integer} column number of data \code{x} imported.
\item The integer value \code{0} or \code{-1} to indicate \code{rownames(x)}
\item One of these strings, to indicate \code{rownames(x)}:
\code{"rownames"}, \code{"rowname"}, or \code{"row.names"}.
\item The default is to use the first column in the imported data \code{x}.
}}

\item{row_annotation_columns}{defines columns to retain in \code{rowData()}
which have non-measurement data associated with each row.
One of the following:
\itemize{
\item \code{character} vector with one or more \code{colnames()}
\item \code{integer} vector with one or more column numbers
}}

\item{curation_txt}{either \code{data.frame} or \code{character} file path
to tab- or comma-delimited file.
The first column should match the
column headers after importing data, \code{colData(se)}.
Subsequent columns contain associated sample annotations.
For Nanostring data, the Nanostring sample annotations will already
be associated with the \code{colData(se)}, and \code{colnames(curation_df)}
will overwrite any that already exist.
\itemize{
\item Pro tip: The first column in \code{curation_txt} should contain \code{'.'}
instead of punctuation/whitespace, to improve pattern matching
filenames where the punctuation characters may have been modified
during processing.
\item Note also that when \code{curation_txt} is supplied, samples in \code{se}
will be subset to include only those samples that match \code{curation_txt},
and in the order they appear in the \code{curation_txt} file.
This behavior allows the \code{curation_txt} to be used to define
the appropriate experimental ordering, which by default also
defines downstream control factor levels for statistical contrasts.
The first factor level is used as the control value in those
contrasts.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to internal functions
for example \code{data.table::fread()}.}
}
\value{
\code{SummarizedExperiment}
}
\description{
Import generic omics tab- or comma-delimited data file
}
\examples{
x <- matrix(letters[1:9], ncol=3);
rownames(x) <- LETTERS[1:3];
colnames(x) <- letters[1:3];
x
import_omics_data(x)

x <- matrix(1:9, ncol=3);
rownames(x) <- LETTERS[1:3];
colnames(x) <- paste0(letters[1:3], "_rep1");
x

curation_txt <- data.frame(Pattern=LETTERS[1:3], Group="A", Batch="B")
se <- import_omics_data(x, curation_txt=curation_txt)
SummarizedExperiment::assays(se)[[1]]
SummarizedExperiment::colData(se)
SummarizedExperiment::rowData(se)

se2 <- import_omics_data(x, curation_txt=curation_txt[c(2, 1, 3), ])
SummarizedExperiment::assays(se2)[[1]]
SummarizedExperiment::colData(se2)

}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_lipotype_csv}()},
\code{\link{import_metabolomics_niehs}()},
\code{\link{import_nanostring_csv}()},
\code{\link{import_nanostring_rcc}()},
\code{\link{import_nanostring_rlf}()},
\code{\link{import_proteomics_PD}()},
\code{\link{import_proteomics_mascot}()},
\code{\link{import_salmon_quant}()},
\code{\link{process_metab_compounds_file}()}
}
\concept{jam import functions}
