% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-proteomics.R
\name{import_proteomics_PD}
\alias{import_proteomics_PD}
\title{Import data from Proteomics Discoverer}
\usage{
import_proteomics_PD(
  xlsx,
  sheet = 1,
  import_types = c("protein", "peptide"),
  ann_lib = c("org.Hs.eg.db"),
  curation_txt = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{xlsx}{\code{character} path to an Excel \code{.xlsx} file as exported
from Proteomics Discoverer software.}
}
\description{
Import data from Proteomics Discoverer
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_nanostring_rcc}()}
}
\concept{jam import functions}
