% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-proteomics.R
\name{import_proteomics_PD}
\alias{import_proteomics_PD}
\title{Import data from Proteomics Discoverer}
\usage{
import_proteomics_PD(
  xlsx,
  sheet = 1,
  import_types = c("protein", "peptide"),
  ann_lib = c("org.Hs.eg.db"),
  curation_txt = NULL,
  remove_duplicate_peptides = TRUE,
  accession_from = NULL,
  accession_to = NULL,
  xref_df = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{xlsx}{\code{character} path to an Excel \code{.xlsx} file as exported
from Proteomics Discoverer software.}

\item{sheet}{\code{integer} or \code{character} used as index or direct character
match with sheet name obtained with \code{openxlsx::getSheetNames(xlsx)}.}

\item{import_types}{\code{character} indicating which type or types of
PD data to import.}

\item{ann_lib}{\code{character} passed to \code{genejam::freshenGenes3()}, see
documentation for alternate methods of passing one or more annotation
libraries.}

\item{curation_txt}{\code{data.frame} whose first column should match the
sample column headers found in the PD abundance columns, and
subsequent columns contain associated sample annotations.
If \code{curation_txt} is not supplied, then values will be split into
columns by \verb{_} underscore or \code{" "} whitespace characters.}

\item{remove_duplicate_peptides}{\code{logical} indicating whether to remove
rows with duplicate sequence-PTM combinations, which can occur when
upstream PD is splitting the same measurement results across multiple
annotation rows. Removing duplicate rows will retain the first
non-duplicated entry in \code{"SeqPTM"} which is composed of the peptide
sequence, and shortened post-translational modification in \code{"PTM"}.}

\item{accession_from, accession_to}{\code{character} vectors, that help manual
curation from one accession number to another, intended when an
accession number is not recognized by the Bioconductor annotation
library, and a newer accession would be recognized. No gene left
behind.}

\item{xref_df}{\code{data.frame} that contains accession numbers in the
first column, and annotation columns in additional columns, specifically
using \verb{"SYMBOL", "ENTREZID", "GENENAME"} as replacements for
output from \code{genejam::freshenGenes3()}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Import data from Proteomics Discoverer
}
\details{
This function is intended to provide a series of steps that import
proteomics abundance data produced by Proteomics Discoverer (PD)
and return a \code{SummarizedExperiment} object ready for downstream
analysis.
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_lipotype_csv}()},
\code{\link{import_nanostring_csv}()},
\code{\link{import_nanostring_rcc}()},
\code{\link{import_salmon_quant}()}
}
\concept{jam import functions}
