% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import-salmonquant.R
\name{import_salmon_quant}
\alias{import_salmon_quant}
\title{Import Salmon quant.sf files to SummarizedExperiment}
\usage{
import_salmon_quant(
  salmonOut_paths,
  import_types = c("tx", "gene"),
  gtf = NULL,
  tx2gene = NULL,
  curation_txt = NULL,
  tx_colname = "transcript_id",
  gene_colname = "gene_name",
  geneFeatureType = "exon",
  txFeatureType = "exon",
  countsFromAbundance = "lengthScaledTPM",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{salmonOut_paths}{\code{character} vectors to each individual folder
that contains the \code{"quant.sf"} output file for Salmon.}

\item{import_types}{\code{character} indicating which type or types of
data to return:
\itemize{
\item \code{tx}: transcript quantitation
\item \code{gene}: gene quantitation after calling \code{tximport::summarizeToGene()}
}}

\item{gtf}{\code{character} path to a GTF file, used only when \code{tx2gene}
is not supplied. When used, \code{splicejam::makeTx2geneFromGtf()} is
called to create a \code{data.frame} object \code{tx2gene}.}

\item{tx2gene}{\code{character} path to file, or \code{data.frame} with at
least two columns matching \code{tx_colname} and \code{gene_colname} below.
When supplied, the \code{gtf} argument is ignored, unless the file
path is not accessible, or the data is not \code{data.frame}.}

\item{curation_txt}{\code{data.frame} whose first column should match the
sample column headers found in the PD abundance columns, and
subsequent columns contain associated sample annotations.
If \code{curation_txt} is not supplied, then values will be split into
columns by \verb{_} underscore or \code{" "} whitespace characters.}

\item{tx_colname, gene_colname}{\code{character} strings indicating colnames
in \code{tx2gene} that should be used.
\itemize{
\item \code{tx_colname} represents unique identifier for each transcript,
usually \code{"transcript_id"}.
\item \code{gene_colname} represents a gene label associated with gene
summarized expression values, typically \code{"gene_name"}.
}}

\item{geneFeatureType, txFeatureType}{\code{character} arguments passed to
\code{splicejam::makeTx2geneFromGtf()} only when supplying argument
\code{gtf} with a path to a GTF file.}

\item{countsFromAbundance}{\code{character} string passed to
\code{tximport::summarizeToGene()} to define the method for calculating
abundance.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to supporting functions.}
}
\description{
Import Salmon quant.sf files to SummarizedExperiment
}
\seealso{
Other jam import functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{deepTools_matrix2nmat}()},
\code{\link{frequency_matrix2nmat}()},
\code{\link{import_lipotype_csv}()},
\code{\link{import_nanostring_csv}()},
\code{\link{import_nanostring_rcc}()},
\code{\link{import_proteomics_PD}()}
}
\concept{jam import functions}
