% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ucsctracks.R
\name{make_ucsc_trackname}
\alias{make_ucsc_trackname}
\title{Make valid UCSC track name from character string}
\usage{
make_ucsc_trackname(x)
}
\arguments{
\item{x}{character vector.

@export}
}
\value{
character vector modified to remove invalid characters.
}
\description{
Make valid UCSC track name from character string
}
\details{
This function takes a character vector, and removes characters
which are considered invalid for a UCSC track name: whitespace
and non-ASCII characters.
}
\seealso{
Other jam ucsc browser functions: 
\code{\link{assign_track_defaults}()},
\code{\link{get_track_defaults}()},
\code{\link{parse_ucsc_gokey}()}
}
\concept{jam ucsc browser functions}
