% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-coverhm.R
\name{nmathm_row_order}
\alias{nmathm_row_order}
\title{Row order for nmatlist coverage heatmap}
\usage{
nmathm_row_order(HM, ...)
}
\arguments{
\item{HM}{\code{list} output from \code{nmatlist2heatmaps()} that
is expected to contain at least one list element
named \code{"HM_drawn"} or \code{"draw"}, with \code{"HM_drawn"}
used first if present. Either object is also expected
to contain slotName ht_list which represents the
HeatmapList. Among the heatmaps in the HeatmapList,
the first with class \code{"Heatmap"} that is not named
"cluster" will be used, since "cluster" is the name
of the heatmap used to represent k-means or other
clustering output to partition rows.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of rownames, where each list represents one
\code{row_slice} used in the heatmap, as defined by
\code{nmatlist2heatmaps()} arguments \code{partition} or
\code{k_clusters}.
}
\description{
Row order for nmatlist coverage heatmap
}
\details{
This function is a simple wrapper function to return the
row order for the output of \code{nmatlist2heatmaps()}.
It traverses the \code{ComplexHeatmap::Heatmap} object,
including optional row slices when the rows are
partitioned. It returns the actual character
vector of rownames, optionally split into a \code{list}.

When the heatmap is fully drawn, for example with
\code{nmatlist2heatmaps(..., do_plot=TRUE)}, the output
object includes an element \code{"HM_drawn"} which also contains
the row order as displayed in that heatmap. Otherwise,
the output contains element \code{"draw"} with the heatmap
that would be drawn by \code{ComplexHeatmap::draw()}. In
this case the row order has not yet been defined,
however this function will evaluate the relevant
criteria to determine the row order that would be
rendered. And note that this process takes slightly
more time, but much less time than rendering the
entire set of heatmaps.
}
