% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-import.R
\name{nmatlist2heatmaps}
\alias{nmatlist2heatmaps}
\title{Make multiple coverage heatmaps}
\usage{
nmatlist2heatmaps(nmatlist, k_clusters = 0, k_subset = NULL,
  k_colors = NULL, k_width = unit(3, "mm"), partition = NULL,
  rows = NULL, nmat_colors = NULL, main_heatmap = 1, hm_nrow = 1,
  transform = jamba::log2signed, lens = -2, seed = 123,
  use_raster = TRUE, do_plot = TRUE, return_type = c("heatmaplist",
  "grid"), verbose = FALSE, ...)
}
\arguments{
\item{nmatlist}{\code{list} containing \code{normalizedMatrix} objects,
usually the output from \code{coverage_matrix2nmat()}.}

\item{k_clusters}{integer number of k-means clusters to
use to partition each heatmap. Use \code{0} or \code{NULL} for
no clustering.}

\item{k_subset}{integer vector of k-means clusters to
retain. Often one cluster contains mostly empty
values, and can be removed using this mechanism.}

\item{k_colors}{vector of R colors, or \code{NULL} to use
the output of \code{colorjam::rainbowJam(k_clusters)}.}

\item{k_width}{unit width of the k-means cluster color
bar, used with \code{k_clusters}.}

\item{partition}{vector used to split rows of each
matrix in \code{nmatlist}, named by rownames. This
value is ignored when \code{k_clusters} is supplied.}

\item{rows}{vector of \code{rownames} or integer vector with
index of rows to keep from each matrix in \code{nmatlist}.}

\item{nmat_colors}{named character vector of R colors,
to colorize each heatmap. When \code{NULL} then
\code{colorjam::rainbowJam()} is used to create colors
for each heatmap panel.}

\item{main_heatmap}{integer index referring to the
entry in \code{nmatlist} to use for clustering and row
ordering.}

\item{hm_nrow}{integer number of rows used to display
the heatmap panels.}

\item{transform}{\code{function} used to transform numeric
values in each entry in \code{nmatlist}.}

\item{lens}{numeric value used to scale each heatmap
color ramp, using \code{getColorRamp()}.}

\item{seed}{numeric value used with \code{set.seed()} to
set the random seed. Set to \code{NULL} to avoid running
\code{set.seed()}.}

\item{use_raster}{logical indicating whether to create heatmaps
using raster resizing, almost always recommended \code{TRUE}.}

\item{do_plot}{logical indicating whether to draw the heatmaps,
where \code{FALSE} will return the data used to create heatmaps
without actually drawing the heatmaps.}

\item{return_type}{character string indicating the type of
data to return: \code{"heatmaplist"} returns the list of heatmaps,
which can separately be arranged together using
\code{ComplexHeatmap::draw()} or \code{grid::grid.draw()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are sent to \code{colorjam::rainbowJam()}.}
}
\description{
Make multiple coverage heatmaps
}
\details{
This function takes a list of \code{normalizedMatrix} objects,
usually the output of \code{coverage_matrix2nmat()}, and
produces multiple heatmaps using
\code{EnrichedHeatmap}.

This function is intended to be a convenient wrapper to
help keep each data matrix in order, to apply consistent
clustering and filtering across all data matrices,
and to enable optional multi-row heatmap layout.
}
\examples{
## There is a small example file to use for testing
cov_file1 <- system.file("data", "tss_coverage.matrix", package="platjam");
cov_file2 <- system.file("data", "h3k4me1_coverage.matrix", package="platjam");
cov_files <- c(cov_file1, cov_file2);
names(cov_files) <- gsub("[.]matrix",
   "",
   basename(cov_files));
nmatlist <- lapply(cov_files, coverage_matrix2nmat);
nmatlist2heatmaps(nmatlist);

# k-means clusters
nmatlist2heatmaps(nmatlist, k_clusters=4);

# multiple rows
nmatlist2heatmaps(nmatlist, k_clusters=4, hm_nrow=2);

}
\seealso{
Other jam genome functions: \code{\link{coverage_matrix2nmat}}
}
\concept{jam genome functions}
