% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverjam-recenter-nmatlist.R
\name{nmatlist_summary}
\alias{nmatlist_summary}
\title{Quick summary of a list of normalizedMatrix data}
\usage{
nmatlist_summary(nmatlist, ...)
}
\arguments{
\item{nmatlist}{\code{list} of \code{normalizedMatrix} objects}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of \code{data.frame} with columns:
\itemize{
\item \code{"row"} with rownames,
\item \code{"summit_name"} when \code{recenter_nmatlist()} has been applied,
\item \code{"restrand"} when \code{restrand_nmatlist()} has been applied.
}
}
\description{
Quick summary of a list of normalizedMatrix data
}
\seealso{
Other jam utility functions: 
\code{\link{cardinality}()},
\code{\link{color_complement}()},
\code{\link{convert_PD_df_to_SE}()},
\code{\link{convert_imputed_assays_to_na}()},
\code{\link{curate_se_colData}()},
\code{\link{curate_to_df_by_pattern}()},
\code{\link{design2layout}()},
\code{\link{get_numeric_transform}()},
\code{\link{handle_df_args}()},
\code{\link{merge_proteomics_se}()},
\code{\link{nmat_summary}()},
\code{\link{rmd_tab_iterator}()},
\code{\link{rowNormScale}()},
\code{\link{summit_from_vector}()}
}
\concept{jam utility functions}
