% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{parse_salmon_flenfile}
\alias{parse_salmon_flenfile}
\title{Parse Salmon fragment length file}
\usage{
parse_salmon_flenfile(
  x,
  k = 7,
  do_plot = FALSE,
  retcolors = c("red", "gold", "dodgerblue", "grey"),
  probs = c(0.25, 0.75),
  ...
)
}
\arguments{
\item{x}{\code{character} path to file usually named \code{"flenDist.txt"}.}

\item{k}{\code{numeric} used with \code{caTools::runmean()} to define the
width for smoothing a running mean, helpful to de-noise
the fragment length profile.}

\item{do_plot}{\code{logical} indicating whether to plot the data and
resulting values.}

\item{retcolors}{\code{character} vector with three colors used when
\code{do_plot=TRUE}: mean, median, mode.}

\item{...}{additional arguments are passed to \code{plot()} when \code{do_plot=TRUE}.}
}
\value{
\code{numeric} vector with summary values:
\itemize{
\item mean: the weighted mean across the fragment length distribution
\item median: the position with 50\% cumulative density at or below
\item mode: the position with the highest density
\item each value in \code{probs} representing percentiles
}

value representing the weighted mean across the
fragment lengths reported, typically values from 1 to 1000.
}
\description{
Parse Salmon fragment length file
}
\seealso{
Other jam nextgen sequence functions: 
\code{\link{get_salmon_meta}()},
\code{\link{get_salmon_root}()},
\code{\link{save_salmon_qc_xlsx}()}
}
\concept{jam nextgen sequence functions}
