% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-numeric-to-colorfn.R
\name{print_color_list}
\alias{print_color_list}
\title{Print color list of color vectors or color functions}
\usage{
print_color_list(x, do_print = TRUE, colorized = TRUE, ...)
}
\arguments{
\item{x}{\code{list} of \code{character} color vectors, or \code{function} defined by
\code{circlize::colorRamp2()} which encodes rgb colors and breaks
in attributes \code{"colors"} and \code{"breaks"} respectively.}

\item{do_print}{\code{logical} indicating whether to print output.
The \code{list} is returned invisibly.}

\item{colorized}{\code{logical} indicating whether to print output using
\code{jamba::printDebug()} with colorized output.}

\item{...}{additional arguments are ignored.}
}
\description{
Print color list of color vectors or color functions
}
\seealso{
Other jam color functions: 
\code{\link{assign_numeric_colors}()},
\code{\link{design2colors}()},
\code{\link{df_to_numcolors}()},
\code{\link{mean_hue}()},
\code{\link{quick_complement_color}()}
}
\concept{jam color functions}
