% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-quick-complement-color.R
\name{quick_complement_color}
\alias{quick_complement_color}
\title{Quick color complement function}
\usage{
quick_complement_color(
  x,
  space = c("hls", "hcl"),
  preset = "ryb",
  do_plot = FALSE,
  ...
)
}
\value{
\code{character} vector of colors
}
\description{
Quick color complement function
}
\examples{
x <- "dodgerblue";
jamba::showColors(list(x=x,
   `quick_complement_color(x)`=quick_complement_color(x),
   `colorjam::color_complement(x, preset="ryb")`=colorjam::color_complement(x, preset="ryb")))

x <- rainbow_hcl(10);
jamba::showColors(list(x=x,
   `quick_complement_color(x)`=quick_complement_color(x),
   `colorjam::color_complement(x, preset="ryb")`=colorjam::color_complement(x, preset="ryb")))

}
\seealso{
Other jam color functions: 
\code{\link{assign_numeric_colors}()},
\code{\link{design2colors}()},
\code{\link{df_to_numcolors}()},
\code{\link{mean_hue}()},
\code{\link{print_color_list}()}
}
\concept{jam color functions}
