% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-quick-complement-color.R
\name{quick_complement_color}
\alias{quick_complement_color}
\title{Quick color complement function}
\usage{
quick_complement_color(x, space = c("hls", "hcl"), do_plot = FALSE, ...)
}
\description{
Quick color complement function
}
\examples{
x <- "dodgerblue";
complement_color(x);
complement_color(x, space="hcl");

x <- c("dodgerblue", "firebrick2", "darkorange2", "gold", "aquamarine2", "darkorchid3");
y <- complement_color(x);
complement_color(y$output_hls)

complement_color("green")
complement_color(rainbow_hcl(10))
complement_color(complement_color(rainbow_hcl(10))[[2]])

x <- c("dodgerblue", "firebrick2", "darkorange2", "gold", "aquamarine2", "darkorchid3");
y <- complement_color(x, preset="rgb");
jamba::showColors(lapply(y, dichromat::dichromat))

}
