% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam_rmd_tab_iterator.R
\name{rmd_tab_iterator}
\alias{rmd_tab_iterator}
\title{RMarkdown tab iterator}
\usage{
rmd_tab_iterator(
  tab_list,
  fn_list = NULL,
  base_fn = NULL,
  tab_labels = NULL,
  heading_level = 2,
  verbose = FALSE,
  htmlOut = TRUE,
  envir = NULL,
  ...
)
}
\arguments{
\item{tab_list}{\code{list} of \code{character} or \code{list} objects.
\itemize{
\item each \code{list} entry represents a layer of tabs
\item \code{names(tab_list)} defines the variable name to which each element
in the \code{list} is assigned.
For example \verb{tab_list=list(use_gene=c("ACTB", "GAPDH")} will create
a layer of tabs, with tabs \code{"ACTB"}, and \code{"GAPDH"}. For the \code{"ACTB"}
tab it will assign \code{use_gene <- "ACTB"} for use in the tab.
\item the names of the \code{list} should match \code{names(prep_list)}
\item each \code{character} vector contains the tabs to display
}}

\item{fn_list}{\code{list} of \code{functions} to apply to corresponding
layers of tabs.
\itemize{
\item \code{names(prep_list)} must match \code{names(tab_list)}, and provides
the function to be applied at the specified layer of tabs.
This function should create the tab content, and should not print
the tab header itself.
}}

\item{...}{additional arguments are ignored.}

\item{final_fn}{\code{function} called at the bottom layer of tabs, after
the corresponding function from \code{fn_list} is called, if it exists.
This function is optional, and is intended to be a convenient place
for the core function in the stack, without having to include this
function in the final layer of \code{fn_list}.}
}
\value{
\code{NULL} invisibly, this function is called for the by-product
of printing RMarkdown-compatible output.
}
\description{
RMarkdown tab iterator
}
\details{
This function is intended to help automate the process of generating
tabs in RMarkdown output, particularly when there may be multiple
layers of tabs. The driving example plots a figure at the lowest level
of tabs, using the options defined by the tabs collectively.

While tabs can be created using \code{for} loops or \code{lapply()} methods,
the RMarkdown code quickly becomes complicated and difficult to maintain.

Each layer of tabs is iterated, and a preparatory function can be called,
for example to update data to be used downstream.
Any layer is free to create a figure, table, or text output, however
it is usually recommended to create output at the lowest layer.

Output can be defined in the final \code{fn_list} layer, however it is cleaner
to use the argument \code{base_fn} to generate final output such as figure
or table. The functions in \code{fn_list} should be limited to those required
for updates during the process.
}
\examples{
tab_labels <- list(
   assay_name=c(
      raw="raw data",
      jammanorm_raw="log ratio normalized",
      quantile_raw="quantile normalized"),
   centerby_name=c(
      global="global-centered",
      `1st_control`="centered vs first control")
   )
base_fn <- function(...){
   plot(x=c(0, 3), y=c(0, 3), asp=1, pch=".")
   text(x=c(1, 1.5, 2), y=c(1, 1.5, 2),
      c(paste0("assay_name:\n", assay_name),
      paste0("formatting:\n", formatting),
      paste0("centerby_name:\n", centerby_name)))
}
tab_list <- list(
   assay_name=c("raw", "jammanorm_raw", "quantile_raw"),
   formatting=c("tab-delimited", "RData"),
   centerby_name=c("global", "1st_control"))
rmd_tab_iterator(
   tab_list=tab_list,
   fn_list=NULL,
   base_fn=base_fn,
   tab_labels=tab_labels,
   heading_level=3,
   verbose=FALSE)

}
\seealso{
Other jam utility functions: 
\code{\link{cardinality}()},
\code{\link{color_complement}()},
\code{\link{convert_PD_df_to_SE}()},
\code{\link{convert_imputed_assays_to_na}()},
\code{\link{curate_se_colData}()},
\code{\link{curate_to_df_by_pattern}()},
\code{\link{design2layout}()},
\code{\link{get_numeric_transform}()},
\code{\link{handle_df_args}()},
\code{\link{merge_proteomics_se}()},
\code{\link{rowNormScale}()}
}
\concept{jam utility functions}
