% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-neuro.R
\name{rowNormScale}
\alias{rowNormScale}
\title{Normalize and scale data per row}
\usage{
rowNormScale(x, from = 0, to = 1, naValue = NA,
  singletMethod = "min", col_range = NULL, ...)
}
\description{
Normalize and scale data per row
}
