% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-ngs.R
\name{save_salmon_qc_xlsx}
\alias{save_salmon_qc_xlsx}
\title{Save Salmon QC to Xlsx}
\usage{
save_salmon_qc_xlsx(
  metajsons,
  salmon_qc_xlsx = NULL,
  adj_target_reads = 75000000,
  ...
)
}
\description{
Save Salmon QC to Xlsx
}
\seealso{
Other jam nextgen sequence functions: 
\code{\link{get_salmon_meta}()},
\code{\link{get_salmon_root}()}
}
\concept{jam nextgen sequence functions}
