% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covhm-validate-params.R
\name{validate_heatmap_params}
\alias{validate_heatmap_params}
\title{Validate coverage heatmap params for Rmarkdown}
\usage{
validate_heatmap_params(param, ...)
}
\description{
Validate coverage heatmap params for Rmarkdown (INCOMPLETE)
}
\seealso{
Other jam coverage heatmap functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{get_nmat_ceiling}()},
\code{\link{nmathm_row_order}()},
\code{\link{nmatlist2heatmaps}()},
\code{\link{recenter_nmatlist}()},
\code{\link{restrand_nmatlist}()},
\code{\link{zoom_nmatlist}()},
\code{\link{zoom_nmat}()}
}
\concept{jam coverage heatmap functions}
