% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covhm-validate-params.R
\name{validate_heatmap_params}
\alias{validate_heatmap_params}
\title{Validate coverage heatmap params for Rmarkdown}
\usage{
validate_heatmap_params(param, ...)
}
\description{
Validate coverage heatmap params for Rmarkdown
}
