% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-coverhm.R
\name{zoom_nmat}
\alias{zoom_nmat}
\title{Zoom the x-axis range for a normalizedMatrix coverage data}
\usage{
zoom_nmat(nmat, upstream_length = 500, downstream_length = 500, ...)
}
\description{
Zoom the x-axis range for a normalizedMatrix coverage data
}
\details{
This function is typically called by \code{zoom_nmatlist()} but can
be called on an individual \code{normalizedMatrix} object.
}
