% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-coverhm.R
\name{zoom_nmatlist}
\alias{zoom_nmatlist}
\title{Zoom the x-axis range for a list of normalizedMatrix coverage data}
\usage{
zoom_nmatlist(nmatlist, upstream_length = 500, downstream_length = 500, ...)
}
\arguments{
\item{nmatlist}{\code{list} of \code{normalizedMatrix} objects. Each
\code{normalizedMatrix} is passed to \code{zoom_nmat()}.}

\item{upstream_length, downstream_length}{\code{numeric} vector whose
values are recycled to length \code{length(nmatlist)}. Each value is
passed to \code{zoom_nmat()} so each matrix can be zoomed to independent
ranges.}

\item{...}{additional arguments are passed to \code{zoom_nmat()}.}
}
\description{
Zoom the x-axis range for a list of normalizedMatrix coverage data
}
\details{
This function filters the matrix columns by distance, and updates
important associated attributes:
\itemize{
\item \code{attr(nmat, "upstream_index")} - the column index positions upstream the target region
\item \code{attr(nmat, "downstream_index")} - the column index positions downstream the target region
\item \code{attr(nmat, "target_index")} - the column index positions representing the target region
\item \code{attr(nmat, "extend")} - the genomic distance upstream and downstream the target region
}
}
\seealso{
Other jam coverage heatmap functions: 
\code{\link{coverage_matrix2nmat}()},
\code{\link{get_nmat_ceiling}()},
\code{\link{nmathm_row_order}()},
\code{\link{nmatlist2heatmaps}()},
\code{\link{zoom_nmat}()}
}
\concept{jam coverage heatmap functions}
