% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platjam-coverhm.R
\name{zoom_nmatlist}
\alias{zoom_nmatlist}
\title{Zoom the x-axis range for a list of normalizedMatrix coverage data}
\usage{
zoom_nmatlist(nmatlist, upstream_length = 500, downstream_length = 500, ...)
}
\description{
Zoom the x-axis range for a list of normalizedMatrix coverage data
}
